separate models file into smaller modules
