--- perma_web/perma/models.py
@@ -1,2550 +1,0 @@
-import calendar
-from contextlib import contextmanager
-from datetime import datetime
-from datetime import timezone as tz
-from decimal import Decimal
-import hashlib
-import hmac
-import itertools
-import json
-import logging
-import os
-import random
-import re
-import time
-from urllib.parse import urlparse
-import uuid
-import zipfile
-
-from dateutil.relativedelta import relativedelta
-from django.conf import settings
-import django.contrib.auth.models
-from django.contrib.auth.models import AbstractBaseUser, BaseUserManager, PermissionsMixin
-from django.contrib.postgres.fields import ArrayField, DateTimeRangeField
-from django.contrib.postgres.indexes import GinIndex, GistIndex, OpClass
-from django.core.files.storage import storages
-from django.db import models, transaction
-from django.db.models import Case, Count, Exists, F, JSONField, Max, OuterRef, Q, Sum, When
-from django.db.models.functions import Now, TruncDate, Upper
-from django.db.models.query import QuerySet
-from django.template.defaultfilters import truncatechars
-from django.urls import reverse
-from django.utils import timezone
-from django.utils.functional import cached_property
-from django.views.decorators.debug import sensitive_variables
-from model_utils import FieldTracker
-from psycopg2.extras import DateTimeTZRange
-import requests
-from rest_framework.settings import api_settings
-import simple_history
-from simple_history.models import HistoricalRecords
-import surt
-from taggit.managers import TaggableManager
-from taggit.models import CommonGenericTaggedItemBase, TaggedItemBase
-from tree_queries.models import TreeNode
-from tree_queries.query import TreeQuerySet
-
-from perma.exceptions import InvalidTransmissionException, PermaPaymentsCommunicationException
-from perma.utils import (
-    first_day_of_next_month,
-    pp_date_from_post,
-    prep_for_perma_payments,
-    preserve_perma_wacz,
-    process_perma_payments_transmission,
-    protocol,
-    remove_control_characters,
-    today_next_year,
-    tz_datetime,
-)
-
-logger = logging.getLogger(__name__)
-
-### CONSTANTS
-ACTIVE_SUBSCRIPTION_STATUSES = ['Current', 'Cancellation Requested']
-PROBLEM_SUBSCRIPTION_STATUSES = ['Hold']
-
-FIELDS_REQUIRED_FROM_PERMA_PAYMENTS = {
-    'get_subscription': [
-        'customer_pk',
-        'customer_type',
-        'subscription',
-        'purchases'
-    ],
-    'get_purchase_history': [
-        'customer_pk',
-        'customer_type',
-        'purchase_history'
-    ]
-}
-
-CUSTOMER_TYPE_MAP = {
-    'LinkUser': 'Individual',
-    'Registrar': 'Registrar'
-}
-
-
-### HELPERS ###
-
-# functions
-def link_count_in_time_period(links, start_time=None, end_time=None):
-    if start_time and end_time and (start_time > end_time):
-        raise ValueError("specified end time is earlier than specified start time")
-    elif start_time and end_time and (start_time == end_time):
-        links = links.filter(creation_timestamp=start_time)
-    else:
-        if start_time:
-            links = links.filter(creation_timestamp__gte=start_time)
-        if end_time:
-            links = links.filter(creation_timestamp__lte=end_time)
-    return links.count()
-
-def most_active_org_in_time_period(organizations, start_time=None, end_time=None):
-    if start_time and end_time and (start_time > end_time):
-        raise ValueError("specified end time is earlier than specified start time")
-    # unlike 'link_count_in_time_period', no special behavior required
-    # if start_time = end_time here. the end result is the same
-    else:
-        if start_time:
-            organizations = organizations.filter(links__creation_timestamp__gte=start_time)
-        if end_time:
-            organizations = organizations.filter(links__creation_timestamp__lte=end_time)
-        return organizations\
-            .annotate(num_links=Count('links'))\
-            .exclude(num_links=0)\
-            .order_by('-num_links')\
-            .first()
-
-def subscription_is_active(subscription):
-    return subscription and (
-        subscription['status'] in ACTIVE_SUBSCRIPTION_STATUSES or (
-            subscription['status'] == "Canceled" and
-            subscription['paid_through'] and
-            subscription['paid_through'] >= timezone.now()
-        )
-    )
-
-def subscription_has_problem(subscription):
-    return subscription and subscription['status'] in PROBLEM_SUBSCRIPTION_STATUSES
-
-
-# classes
-
-class DeletableManager(models.Manager):
-    """
-        Manager that excludes results where user_deleted=True by default.
-    """
-    def get_queryset(self):
-        # exclude deleted entries by default
-        return super(DeletableManager, self).get_queryset().filter(user_deleted=False)
-
-    def all_with_deleted(self):
-        return super(DeletableManager, self).get_queryset()
-
-
-class DeletableModel(models.Model):
-    """
-        Abstract base class that lets a model track deletion.
-    """
-    user_deleted = models.BooleanField(default=False, verbose_name="Deleted by user")
-    user_deleted_timestamp = models.DateTimeField(null=True, blank=True)
-
-    class Meta:
-        abstract = True
-
-    def safe_delete(self):
-        self.user_deleted = True
-        self.user_deleted_timestamp = timezone.now()
-
-# django-taggit assumes the model being tagged has an integer primary key.
-# per http://django-taggit.readthedocs.io/en/latest/custom_tagging.html,
-# tag "through" this class if your model has a string as primary key.
-# tags = TaggableManager(through=GenericStringTaggedItem)
-# (copied straight from their docs)
-class GenericStringTaggedItem(CommonGenericTaggedItemBase, TaggedItemBase):
-    object_id = models.CharField(max_length=50, db_index=True)
-
-
-class CustomerModel(models.Model):
-    """
-        Abstract base class that lets a model upgrade to a paid account.
-    """
-    class Meta:
-        abstract = True
-
-    nonpaying = models.BooleanField(default=False, help_text="Whether this customer qualifies for a free account.")
-    in_trial = models.BooleanField(default=True, help_text="Is this customer in their trial period?")
-    base_rate =  models.DecimalField(
-        max_digits=19,
-        decimal_places=2,
-        default=Decimal(settings.DEFAULT_BASE_RATE),
-        help_text="Base rate for calculating subscription cost."
-    )
-    cached_subscription_started = models.DateTimeField(
-        null=True,
-        blank=True,
-        help_text="Used to help calculate how many links have been created against a paying customer's link limit."
-    )
-    cached_subscription_status = models.CharField(
-        max_length=50,
-        null=True,
-        blank=True,
-        help_text="The last known status of customer's paid subscription, from Perma Payments"
-    )
-    cached_paid_through = models.DateTimeField(
-        null=True,
-        blank=True
-    )
-    cached_subscription_rate = models.DecimalField(
-        max_digits=19,
-        decimal_places=2,
-        null=True,
-        blank=True,
-        help_text="Local record of what the customer last paid."
-    )
-    unlimited = models.BooleanField(default=False, help_text="If unlimited, link_limit and related fields are ignored.")
-    link_limit = models.IntegerField(default=settings.DEFAULT_CREATE_LIMIT)
-    link_limit_period = models.CharField(max_length=8, default=settings.DEFAULT_CREATE_LIMIT_PERIOD, choices=(('once','once'),('monthly','monthly'),('annually','annually')))
-    bonus_links = models.PositiveIntegerField(blank=True, null=True)
-
-    @cached_property
-    def customer_type(self):
-        return CUSTOMER_TYPE_MAP[type(self).__name__]
-
-    @sensitive_variables()
-    def get_purchase_history(self):
-        if self.nonpaying:
-            return None
-
-        try:
-            r = requests.post(
-                settings.PURCHASE_HISTORY_URL,
-                timeout=settings.PERMA_PAYMENTS_TIMEOUT,
-                data={
-                    'encrypted_data': prep_for_perma_payments({
-                        'timestamp': datetime.utcnow().timestamp(),
-                        'customer_pk':  self.pk,
-                        'customer_type': self.customer_type
-                    })
-                }
-            )
-            assert r.ok, r.status_code
-        except (requests.RequestException, AssertionError) as e:
-            msg = f"Communication with Perma-Payments failed: {e}"
-            if settings.PERMA_PAYMENTS_IN_MAINTENANCE:
-                logger.info(msg)
-            else:
-                logger.error(msg)
-            raise PermaPaymentsCommunicationException(msg)
-
-        post_data = process_perma_payments_transmission(r.json(), FIELDS_REQUIRED_FROM_PERMA_PAYMENTS['get_purchase_history'])
-
-        if post_data['customer_pk'] != self.pk or post_data['customer_type'] != self.customer_type:
-            msg = "Unexpected response from Perma-Payments."
-            logger.error(msg)
-            raise InvalidTransmissionException(msg)
-
-        return {
-            'purchases': [
-                {
-                    'link_quantity': item['link_quantity'],
-                    'date': pp_date_from_post(item['date']),
-                    'reference_number': item['reference_number']
-                } for item in post_data['purchase_history']
-            ],
-            'total_links': sum(int(purchase['link_quantity']) for purchase in post_data['purchase_history'])
-        }
-
-
-    @sensitive_variables()
-    def get_subscription(self):
-        if self.nonpaying:
-            return None
-
-        try:
-            r = requests.post(
-                settings.SUBSCRIPTION_STATUS_URL,
-                timeout=settings.PERMA_PAYMENTS_TIMEOUT,
-                data={
-                    'encrypted_data': prep_for_perma_payments({
-                        'timestamp': datetime.utcnow().timestamp(),
-                        'customer_pk':  self.pk,
-                        'customer_type': self.customer_type
-                    })
-                }
-            )
-            assert r.ok, r.status_code
-        except (requests.RequestException, AssertionError) as e:
-            msg = f"Communication with Perma-Payments failed: {e}"
-            if settings.PERMA_PAYMENTS_IN_MAINTENANCE:
-                logger.info(msg)
-            else:
-                logger.error(msg)
-            raise PermaPaymentsCommunicationException(msg)
-
-        post_data = process_perma_payments_transmission(r.json(), FIELDS_REQUIRED_FROM_PERMA_PAYMENTS['get_subscription'])
-
-        if post_data['customer_pk'] != self.pk or post_data['customer_type'] != self.customer_type:
-            msg = "Unexpected response from Perma-Payments."
-            logger.error(msg)
-            raise InvalidTransmissionException(msg)
-
-        #
-        # First, credit the user for any bonus links they have purchased.
-        #
-        if post_data['purchases']:
-            self.credit_for_purchased_links(post_data['purchases'])
-
-        #
-        # Then, handle subscription-related concerns
-        #
-        if post_data['subscription'] is None:
-            if self.cached_subscription_started:
-                # reset this, so that link counts work properly if the customer
-                # purchases a new subscription in the future
-                self.cached_subscription_started = None
-                self.save(update_fields=['cached_subscription_started'])
-                self.refresh_from_db()
-            return None
-
-        # Alert Perma that this user is no longer in their trial period.
-        # Store the subscription status locally, for use if Perma Payments is unavailable
-        # and update local link limit and rate to match Perma Payments' records
-        subscription_change_effective = pp_date_from_post(post_data['subscription']['link_limit_effective_timestamp'])
-        self.in_trial = False
-        if not self.cached_subscription_started:
-            self.cached_subscription_started = subscription_change_effective
-        self.cached_subscription_status = post_data['subscription']['status']
-        self.cached_paid_through = pp_date_from_post(post_data['subscription']['paid_through'])
-
-        pending_change = None
-        if subscription_change_effective <= timezone.now():
-            self.link_limit_period = post_data['subscription']['frequency']
-            self.cached_subscription_rate = Decimal(post_data['subscription']['rate'])
-            if post_data['subscription']['link_limit'] == 'unlimited':
-                self.unlimited = True
-            else:
-                self.unlimited = False
-                self.link_limit = int(post_data['subscription']['link_limit'])
-        else:
-            pending_change = {
-                'rate': post_data['subscription']['rate'],
-                'link_limit': post_data['subscription']['link_limit'],
-                'effective': subscription_change_effective
-            }
-        self.save(update_fields=['in_trial', 'cached_subscription_started', 'cached_subscription_status', 'cached_paid_through', 'cached_subscription_rate', 'unlimited', 'link_limit', 'link_limit_period'])
-        self.refresh_from_db()
-
-        return {
-            'status': self.cached_subscription_status,
-            'frequency': self.link_limit_period,
-            'paid_through': self.cached_paid_through,
-            'rate': str(self.cached_subscription_rate),
-            'link_limit': 'unlimited' if self.unlimited else str(self.link_limit),
-            'pending_change': pending_change,
-            'reference_number': post_data['subscription']['reference_number'],
-        }
-
-    def annotate_tier(self, tier, current_subscription, now, next_month, next_year):
-        '''
-        Mutates the passed-in tier dictionary, adding time- and subscription-specific details.
-        '''
-
-        # Calculate when, after today, the customer will/should next be charged.
-        # Calculate what fraction of the current subscription period remains,
-        # to use when determining how much to charge them today.
-        if tier['period'] == 'monthly':
-            # monthly subscriptions are paid on the first of the next month
-            next_payment = next_month
-            days_in_month = calendar.monthrange(now.year, now.month)[1]
-            prorated_ratio = Decimal((next_payment - now).days / days_in_month)
-        elif tier['period'] == 'annually':
-            # annual subscriptions are paid on the anniversary of their creation
-            if current_subscription:
-                # n.b. these values are nonsensical if the current subscription is not active.
-                # there is no good answer in that case.... so updating a non-active
-                # subscription is forbidden below. continuing to calculate the nonsensical values
-                # for these fields since.... that at least avoids type errors.
-                next_payment = current_subscription['paid_through']
-                prorated_ratio  = Decimal((next_payment - now).days / 365)  # ignore leap year
-            else:
-                next_payment = next_year
-                prorated_ratio  = Decimal(1)
-        else:
-            raise NotImplementedError(f'Paid "{tier["frequency"]}" tiers not yet supported')
-
-        # Customers without subscriptions may upgrade to any tier.
-        #
-        # Customers with existing non-active subscriptions may not upgrade or downgrade.
-        #
-        # Customers with existing active subscriptions may upgrade/downgrade
-        # to another tier with the same link limit period/payment frequency.
-        #
-        # Upgrades are effective immediately. Today, customers should be
-        # charged the prorated cost of the difference between their current
-        # subscription tier and tier they are upgrading to.
-        #
-        # Downgrades are effective the next time their subscription renews.
-        # The current subscription period will not be affected: customers
-        # should not be charged today.
-        #
-        # If a customer has already scheduled a downgrade for the next
-        # subscription period, all tiers should be unavailable;
-        # the cancellation of scheduled downgrades is handled elsewhere.
-        tier_rate = self.base_rate * Decimal(tier['rate_ratio'])
-
-        if not current_subscription:
-            tier_type = 'upgrade'
-            todays_charge = prorated_ratio * tier_rate
-        elif not current_subscription['status'] == 'Current' \
-             or tier['period'] != current_subscription['frequency'] \
-             or current_subscription.get('pending_change'):
-            tier_type = 'unavailable'
-            todays_charge = Decimal(0)
-        else:
-            current_limit = float('Inf') if current_subscription['link_limit'] == 'unlimited' else float(current_subscription['link_limit'])
-            tier_limit = float('Inf') if tier['link_limit'] == 'unlimited' else float(tier['link_limit'])
-            current_rate = Decimal(current_subscription['rate'])
-
-            if tier_rate == current_rate and tier_limit == current_limit:
-                tier_type = 'selected'
-                todays_charge = Decimal(0)
-            elif tier_rate <= current_rate:
-                if tier_limit >= current_limit:
-                    # This means the customer is overpaying, by today's standards.
-                    # We should not let this happen: solve by granting the user
-                    # more links for their money, via the Perma Payments admin,
-                    # when we lower our tier prices.
-                    logger.error(f"{str(self)} is being overcharged subsequent to new Perma subscription tiers.")
-                    tier_type = 'unavailable'
-                    todays_charge = Decimal(0)
-                else:
-                    tier_type = 'downgrade'
-                    todays_charge = Decimal(0)
-            else:
-                if tier_limit <= current_limit:
-                    # This means the customer is underpaying, by today's standards.
-                    # We should not let them upgrade in the normal way.
-                    # If we don't want this to happen, we should work it out via
-                    # the Perma admin, the Perma Payments admin, and/or CyberSource Business Center
-                    tier_type = 'unavailable'
-                    todays_charge = Decimal(0)
-                else:
-                    tier_type = 'upgrade'
-                    todays_charge = prorated_ratio * (tier_rate - current_rate)
-
-        tier.update({
-            'type': tier_type,
-            'link_limit': str(tier['link_limit']),
-            'link_limit_effective_timestamp': now.timestamp() if tier_type == 'upgrade' else next_payment.timestamp(),
-            'todays_charge': "{0:.2f}".format(todays_charge.quantize(Decimal('.01'))),
-            'recurring_amount': "{0:.2f}".format(tier_rate),
-            'recurring_start_date': next_payment.strftime("%Y-%m-%d"),
-            'next_payment': next_payment
-        })
-
-    def get_subscription_info(self, now):
-        timestamp = now.timestamp()
-        next_month = first_day_of_next_month(now)
-        next_year = today_next_year(now)
-        subscription = self.get_subscription()
-
-        tiers = []
-        if subscription and subscription.get('pending_change'):
-            # allow the user to effective cancel the pending change,
-            # reverting to / rescheduling whatever is on record as
-            # their "current" subscription, in Perma
-            required_fields = {
-                'customer_pk': self.pk,
-                'customer_type': self.customer_type,
-                'timestamp': timestamp,
-                'amount': '0.00',
-                'recurring_amount': subscription['rate'],
-                'recurring_frequency': subscription['frequency'],
-                'recurring_start_date': subscription['paid_through'].strftime("%Y-%m-%d"),
-                'link_limit': subscription['link_limit'],
-                'link_limit_effective_timestamp': now.timestamp()
-            }
-            tiers.append({
-                'type': 'cancel_downgrade',
-                'period': subscription['frequency'],
-                'limit': subscription['link_limit'],
-                'rate': subscription['rate'],
-                'next_payment': subscription['paid_through'].strftime("%Y-%m-%d"),
-                'required_fields': required_fields,
-                'encrypted_data': prep_for_perma_payments(required_fields).decode('utf-8')
-            })
-        else:
-            for tier in settings.TIERS[self.customer_type]:
-                self.annotate_tier(tier, subscription, now, next_month, next_year)
-                required_fields = {
-                    'customer_pk': self.pk,
-                    'customer_type': self.customer_type,
-                    'timestamp': timestamp,
-                    'amount': tier['todays_charge'],
-                    'recurring_amount': tier['recurring_amount'],
-                    'recurring_frequency': tier['period'],
-                    'recurring_start_date': tier['recurring_start_date'],
-                    'link_limit': tier['link_limit'],
-                    'link_limit_effective_timestamp': tier['link_limit_effective_timestamp']
-                }
-                tiers.append({
-                    'type': tier['type'],
-                    'period': tier['period'],
-                    'limit': tier['link_limit'],
-                    'rate': tier['recurring_amount'],
-                    'next_payment': tier['next_payment'],
-                    'required_fields': required_fields,
-                    'encrypted_data': prep_for_perma_payments(required_fields).decode('utf-8')
-                })
-
-        return {
-            'customer': self,
-            'subscription': subscription,
-            'tiers': tiers,
-            'can_change_tiers': any(tier['type'] in ['upgrade', 'downgrade', 'cancel_downgrade'] for tier in tiers)
-        }
-
-    def credit_for_purchased_links(self, purchases):
-        credited_link_count = 0
-        for purchase in purchases:
-            try:
-                with transaction.atomic():
-                    link_quantity = int(purchase["link_quantity"])
-                    self.bonus_links = (self.bonus_links or 0) + link_quantity
-                    self.save(update_fields=['bonus_links'])
-                    try:
-                        r = requests.post(
-                            settings.ACKNOWLEDGE_PURCHASE_URL,
-                            timeout=settings.PERMA_PAYMENTS_TIMEOUT,
-                            data={
-                                'encrypted_data': prep_for_perma_payments({
-                                    'timestamp': datetime.utcnow().timestamp(),
-                                    'purchase_pk':  purchase['id']
-                                })
-                            }
-                        )
-                        assert r.ok, r.status_code
-                    except (requests.RequestException, AssertionError) as e:
-                        msg = f"Communication with Perma-Payments failed: {str(e)}"
-                        if settings.PERMA_PAYMENTS_IN_MAINTENANCE:
-                            logger.info(msg)
-                        else:
-                            logger.error(msg)
-                        raise PermaPaymentsCommunicationException(msg)
-                    credited_link_count += link_quantity
-            except PermaPaymentsCommunicationException:
-                # I think we want the function to return even if it fails...
-                # We'll be notified via the error message, and the calling
-                # can do its best to proceed... having failed to credit the user
-                # for their links. (Presumably, the customer will also complain if failure persists.)
-                pass
-        return credited_link_count
-
-    def get_bonus_packages(self):
-        bonus_packages = []
-        for package in settings.BONUS_PACKAGES:
-            required_fields = {
-                'timestamp': datetime.utcnow().timestamp(),
-                'customer_pk':  self.pk,
-                'customer_type': self.customer_type,
-                'amount': package['price'],
-                'link_quantity': package['link_quantity']
-            }
-            bonus_packages.append({
-                'amount': required_fields['amount'],
-                'link_quantity': required_fields['link_quantity'],
-                'unit_cost': float(required_fields['amount']) / int(required_fields['link_quantity']),
-                'encrypted_data': prep_for_perma_payments(required_fields).decode('utf-8')
-            })
-        return bonus_packages
-
-
-    @cached_property
-    def subscription_status(self):
-        try:
-            subscription = self.get_subscription()
-        except PermaPaymentsCommunicationException:
-            subscription = {
-                'status': self.cached_subscription_status,
-                'paid_through': self.cached_paid_through
-            }
-        if subscription_is_active(subscription):
-            return 'active'
-        if subscription_has_problem(subscription):
-            return 'problem'
-        return None
-
-    def link_creation_allowed(self):
-        """
-        Must be implemented by children
-        """
-        raise NotImplementedError
-
-
-### MODELS ###
-
-class RegistrarQuerySet(QuerySet):
-    def approved(self):
-        return self.filter(status="approved")
-
-class Registrar(CustomerModel):
-    """
-    This is a library, a court, a firm, or similar.
-    """
-    name = models.CharField(max_length=400, db_index=True)
-    email = models.EmailField(max_length=254)
-    website = models.URLField(max_length=500)
-    date_created = models.DateTimeField(auto_now_add=True, null=True)
-    status = models.CharField(max_length=20, default='pending', choices=(('pending','pending'),('approved','approved'),('denied','denied')))
-    orgs_private_by_default = models.BooleanField(default=False, help_text="Whether new orgs created for this registrar default to private links.")
-
-    address = models.CharField(max_length=500, blank=True, null=True)
-    manual_sort_order = models.IntegerField(default=0, db_index=True)
-
-    link_count = models.IntegerField(default=0) # A cache of the number of links under this registrars's purview (sum of all associated org links)
-
-    objects = RegistrarQuerySet.as_manager()
-    tracker = FieldTracker()
-    history = HistoricalRecords()
-    tags = TaggableManager(blank=True)
-
-    class Meta:
-        ordering = ['name']
-
-    def __str__(self):
-        return self.name
-
-    def save(self, *args, **kwargs):
-        with self.tracker:
-            super(Registrar, self).save(*args, **kwargs)
-            if self.tracker.has_changed('name'):
-                # Rename top-level sponsored folders if registrar name changes.
-                folders = Folder.objects.filter(sponsored_by=self, parent__is_sponsored_root_folder=True)
-                folders.update(name=self.name)
-
-    def link_count_in_time_period(self, start_time=None, end_time=None):
-        links = Link.objects.filter(organization__registrar=self)
-        return link_count_in_time_period(links, start_time, end_time)
-
-    def link_count_this_year(self):
-        return self.link_count_in_time_period(tz_datetime(timezone.now().year, 1, 1))
-
-    def most_active_org_in_time_period(self, start_time=None, end_time=None):
-        return most_active_org_in_time_period(self.organizations, start_time, end_time)
-
-    def most_active_org_this_year(self):
-        return most_active_org_in_time_period(self.organizations, tz_datetime(timezone.now().year, 1, 1))
-
-    def active_registrar_users(self):
-        return self.users.filter(is_active=True)
-
-    def link_creation_allowed(self):
-        # No logic yet for handling paid Registrar customers with limits:
-        # all paid-up Registrar customers get unlimited links.
-        assert self.unlimited
-        if self.nonpaying:
-            return True
-        if self.id in settings.SPECIAL_UNLIMITED_LINKS_FOR_REGISTRAR:
-            return True
-        return self.subscription_status == 'active'
-
-Registrar._meta.get_field('nonpaying').default = True
-Registrar._meta.get_field('unlimited').default = True
-Registrar._meta.get_field('base_rate').default = Decimal(settings.DEFAULT_BASE_RATE_REGISTRAR)
-
-
-class OrganizationQuerySet(QuerySet):
-    def accessible_to(self, user):
-        qset = self.user_access_filter(user)
-        if qset is None:
-            return self.none()
-        else:
-            return self.filter(qset)
-
-    def user_access_filter(self, user):
-        if user.is_organization_user:
-            return Q(id__in=user.organizations.all())
-        elif user.is_registrar_user():
-            return Q(registrar_id=user.registrar_id)
-        elif user.is_staff:
-            return Q()  # all
-        else:
-            return None
-
-
-OrganizationManager = DeletableManager.from_queryset(OrganizationQuerySet)
-
-
-class Organization(DeletableModel):
-    """
-    This is generally a journal.
-    """
-    name = models.CharField(max_length=400)
-    registrar = models.ForeignKey(Registrar, null=True, related_name="organizations", on_delete=models.CASCADE)
-    shared_folder = models.OneToOneField('Folder', blank=True, null=True, related_name="top_level_for_org", on_delete=models.CASCADE)
-    date_created = models.DateTimeField(auto_now_add=True, null=True)
-    default_to_private = models.BooleanField(default=False)
-    link_count = models.IntegerField(default=0) # A cache of the number of links under this org's purview
-
-    objects = OrganizationManager()
-    tracker = FieldTracker()
-    history = HistoricalRecords()
-
-    class Meta:
-        indexes = [
-            models.Index(fields=['user_deleted']),
-        ]
-
-    def save(self, *args, **kwargs):
-
-        with transaction.atomic():
-
-            if not self.pk:
-                self.default_to_private = self.registrar.orgs_private_by_default
-
-            with self.tracker:
-                # Save here, so we have a PK if we need it below, to create the shared folder
-                super().save(*args, **kwargs)
-
-                if not self.shared_folder_id:
-                    # Create a top-level folder for this org
-                    shared_folder = Folder.objects.create(
-                        name=self.name,
-                        organization=self,
-                        is_shared_folder=True
-                    )
-                    self.shared_folder = shared_folder
-                    # Save with super again, instead of plain save,
-                    # so we don't run through our custom logic twice
-                    super().save()
-
-                elif self.tracker.has_changed('name'):
-                    # Rename shared folder if org name changes.
-                    self.shared_folder.name = self.name
-                    self.shared_folder.save()
-
-    def __str__(self):
-        return self.name
-
-    def link_count_in_time_period(self, start_time=None, end_time=None):
-        links = Link.objects.filter(organization=self)
-        return link_count_in_time_period(links, start_time, end_time)
-
-    def link_count_this_year(self):
-        return self.link_count_in_time_period(tz_datetime(timezone.now().year, 1, 1))
-
-    def accessible_to(self, user):
-        if user.is_staff:
-            return True
-        if user.is_registrar_user():
-            return self.registrar_id == user.registrar_id
-        return self.users.filter(pk=user.pk).exists()
-
-
-class Sponsorship(models.Model):
-    registrar = models.ForeignKey(Registrar, on_delete=models.PROTECT, related_name='sponsorships')
-    user = models.ForeignKey('LinkUser', on_delete=models.CASCADE, related_name='sponsorships')
-    status = models.CharField(max_length=10, blank=True, null=True, choices=(('active','Active: user may create links.'), ('inactive', 'Inactive: user may view, but not create, links.')), default='active')
-    status_changed = models.DateTimeField(auto_now=True)
-    created_at = models.DateTimeField(auto_now_add=True)
-    created_by = models.ForeignKey('LinkUser', related_name='created_sponsorships', on_delete=models.PROTECT)
-    expires_at = models.DateTimeField(blank=True, null=True)
-
-    class Meta:
-        constraints = [
-            models.UniqueConstraint(fields=['registrar', 'user'], name='unique_sponsorship'),
-        ]
-        indexes = [
-            models.Index(fields=['status', 'expires_at'])
-        ]
-
-    tracker = FieldTracker()
-
-    def save(self, *args, **kwargs):
-        with transaction.atomic():
-            with self.tracker:
-                super().save(*args, **kwargs)
-                if not self.folders:
-                    self.user.create_sponsored_folder(self.registrar)
-                if self.tracker.has_changed('status'):
-                    self.folders.update(read_only=self.status == 'inactive')
-
-    @property
-    def folders(self):
-        return Folder.objects.filter(owned_by=self.user, sponsored_by=self.registrar)
-
-
-class LinkUserManager(BaseUserManager):
-    def create_user(self, email, registrar, organization, date_joined, first_name, last_name, authorized_by, password=None):
-        """
-        Creates and saves a User with the given email, registrar and password.
-        """
-
-        if not email:
-            raise ValueError('Users must have an email address')
-        if registrar and organization:
-            raise ValueError('Users may not have both registrar and organization affiliations.')
-
-        user = self.model(
-            email=self.normalize_email(email),
-            registrar=registrar,
-            date_joined = date_joined,
-            first_name = first_name,
-            last_name = last_name,
-            authorized_by = authorized_by,
-        )
-
-        user.set_password(password)
-        user.save()
-
-        if organization:
-            user.organizations.add(organization)
-
-        return user
-
-
-# This is a temporary workaround for the problem described in
-# https://github.com/jazzband/django-model-utils/issues/331#issuecomment-478994563
-# where django-model-utils FieldTracker breaks the setter for overridden attributes on abstract base classes
-del AbstractBaseUser.is_active
-
-class LinkUser(CustomerModel, AbstractBaseUser, PermissionsMixin):
-    email = models.EmailField(
-        verbose_name='email address',
-        max_length=255,
-        unique=True,
-        db_index=True,
-        error_messages={'unique': "A user with that email address already exists.",}
-    )
-    raw_email = models.CharField(
-        verbose_name='raw email address',
-        max_length=255,
-        null=True
-    )
-    registrar = models.ForeignKey(Registrar, blank=True, null=True, related_name='users', help_text="If set, this user is a registrar user. This should not be set if org is set!", on_delete=models.CASCADE)
-    pending_registrar = models.ForeignKey(Registrar, blank=True, null=True, related_name='pending_users', on_delete=models.CASCADE)
-    organizations = models.ManyToManyField(Organization, through='UserOrganizationAffiliation', blank=True, related_name='users',
-                                           help_text="If set, this user is an org user. This should not be set if registrar is set!<br><br>"
-                                                     "Note: <b>This list will include deleted orgs of which this user is a member.</b> This is a historical"
-                                                     " record and deleted org memberships cannot be removed.<br><br>"
-                                           )
-    sponsoring_registrars = models.ManyToManyField(
-        Registrar,
-        blank=True,
-        related_name='sponsored_users',
-        through=Sponsorship,
-        through_fields=('user', 'registrar'),
-        help_text="If set, this user is sponsored by a registrar. Any user can be sponsored by any registrar."
-    )
-    is_active = models.BooleanField(default=False)
-    is_confirmed = models.BooleanField(default=False)
-    is_staff = models.BooleanField(default=False)
-    date_joined = models.DateTimeField(auto_now_add=True)
-    first_name = models.CharField(max_length=45, blank=True)
-    last_name = models.CharField(max_length=45, blank=True)
-    root_folder = models.OneToOneField('Folder', blank=True, null=True, on_delete=models.CASCADE)
-    sponsored_root_folder = models.OneToOneField('Folder', blank=True, null=True, on_delete=models.CASCADE, related_name='sponsored_user')
-    requested_account_type = models.CharField(max_length=45, blank=True, null=True)
-    requested_account_note = models.CharField(max_length=45, blank=True, null=True)
-    link_count = models.IntegerField(default=0) # A cache of the number of links created by this user
-    notes = models.TextField(blank=True)
-
-    objects = LinkUserManager()
-    # Don't add an unconfigured FieldTracker() to LinkUser: it breaks last_login https://github.com/harvard-lil/perma/issues/3296
-    # If tracking is required, enumerate the necessary fields using the fields param https://django-model-utils.readthedocs.io/en/latest/utilities.html#tracking-specific-fields
-    # tracker = FieldTracker(fields=[])
-
-    USERNAME_FIELD = 'email'
-    REQUIRED_FIELDS = []
-
-    class Meta:
-        verbose_name = 'User'
-
-        constraints = [
-            # We would like to add a case-insensitive uniqueness constraint on `email`
-            # like the below, but expressions aren't supported in this version of Django.
-            # We are adding the constraint via a SQL migration instead. See 0019_auto_20230424_2134.py
-            # We additionally note that in this version of Django, constraints aren't
-            # checked during model validation. We are not adding a special case-insensitive
-            # validation check, because we downcase email in `clean`: checked against a
-            # known-to-be-downcased column, the check from unique=True is sufficient.
-            # models.UniqueConstraint(Lower('email'), name='unique_lower_email')
-        ]
-
-    def format_email_fields(self):
-        self.raw_email = self.email
-        self.email = self.email.lower()
-
-    def clean(self, *args, **kwargs):
-        self.format_email_fields()
-        super().clean(*args, **kwargs)
-
-    def save(self, *args, **kwargs):
-        if not self.pk and not self.raw_email:
-            # If objects aren't being created via a model form where `clean` has been called,
-            # make sure email is still formatted correctly.
-            self.format_email_fields()
-
-        with transaction.atomic():
-            super().save(*args, **kwargs)
-
-            # make sure root folder is created for each user.
-            if not self.root_folder_id:
-                root_folder = Folder.objects.create(
-                    name='Personal Links',
-                    created_by=self,
-                    is_root_folder=True
-                )
-                self.root_folder = root_folder
-                # Save with super again, instead of plain save,
-                # so we don't run through our custom logic twice
-                super().save()
-
-    def get_full_name(self):
-        """ Use either First Last or first half of email address as user's name. """
-        return f"{self.first_name} {self.last_name}" if self.first_name or self.last_name else self.email.split('@')[0]
-
-    def get_short_name(self):
-        """ Use either First or Last or first half of email address as user's short name. """
-        return self.first_name or self.last_name or self.email.split('@')[0]
-
-    def __str__(self):
-        return self.email
-
-    def top_level_folders(self):
-        """
-            Get top level folders for this user, including personal folder, sponsored folder, and shared folders.
-            Returns a queryset of Folders, with personal/sponsored folders first, followed by org folders sorted by name.
-        """
-        # personal folder first, custom_order = 0
-        folder_ids = [self.root_folder_id]
-        ordering_cases = [When(id=self.root_folder_id, then=0)]
-
-        # sponsored folder second, custom_order = 1
-        if self.sponsored_root_folder_id:
-            folder_ids.append(self.sponsored_root_folder_id)
-            ordering_cases.append(When(id=self.root_folder_id, then=1))
-
-        # then org folders if any, custom_order = 2
-        return Folder.objects.filter(
-            Q(id__in=folder_ids) |
-            Q(id__in=self.get_orgs().values('shared_folder_id'))
-        ).annotate(
-            custom_order=Case(*ordering_cases, default=2)
-        ).order_by('custom_order', 'name')
-
-    def all_folder_trees(self):
-        """
-            Get all folders for this user, including personal folders and shared folders.
-        """
-        return [folder.get_descendants(include_self=True) for folder in self.top_level_folders()]
-
-    def get_orgs(self):
-        """
-            Get organizations in which this user is a member
-        """
-
-        if self.is_organization_user:
-            return self.organizations.all()
-        if self.is_registrar_user():
-            return self.registrar.organizations.all()
-        if self.is_staff:
-            return Organization.objects.all()
-
-        return Organization.objects.none()
-
-    def create_sponsored_folder(self, registrar):
-        with transaction.atomic():
-            if not self.sponsored_root_folder_id:
-                sponsored_root_folder = Folder.objects.create(
-                    name='Sponsored Links',
-                    created_by=self,
-                    is_sponsored_root_folder=True
-                )
-                self.sponsored_root_folder = sponsored_root_folder
-                self.save()
-        Folder.objects.create(
-            name=registrar.name,
-            created_by=self,
-            parent_id=self.sponsored_root_folder_id,
-            sponsored_by=registrar
-        )
-
-    def as_json(self):
-        from api.serializers import LinkUserSerializer  # local import to avoid circular import
-        return json.dumps(LinkUserSerializer(self).data)
-
-    def get_api_key(self):
-        key = None
-        try:
-            key = self.api_key.key
-        except LinkUser.api_key.RelatedObjectDoesNotExist:
-            pass
-        return key
-
-    ### permissions ###
-
-    def has_perm(self, perm, obj=None):
-        """
-            Does the user have a specific permission?
-            Simplest possible answer: Yes, always
-            This is only used by the django admin for is_staff=True users.
-        """
-        return True
-
-    def has_module_perms(self, app_label):
-        """
-            Does the user have permissions to view the app `app_label`?
-            Simplest possible answer: Yes, always
-            This is only used by the django admin for is_staff=True users.
-        """
-        return True
-
-    def shares_scope_with_user(self, other_user):
-        """
-            Does the user share a scope with another user?
-
-            Org users share scope with other members of their orgs.
-            Registrar users share scope with others registrar users from
-               the same registrar, sponsored users, and all members of the registrar's orgs.
-            Admins share scope with all users.
-        """
-        if self.is_organization_user:
-            orgs = other_user.organizations.all() & self.organizations.all()
-            return len(orgs) > 0
-        elif self.is_registrar_user():
-            if self.registrar == other_user.registrar:
-                return True
-            if self.registrar in other_user.sponsoring_registrars.all():
-                return True
-            orgs = other_user.organizations.all() & Organization.objects.filter(registrar=self.registrar)
-            return len(orgs) > 0
-        elif self.is_staff:
-            return True
-        return False
-
-    def is_individual(self):
-        """ Is the user a regular, individual user? """
-        return bool(not self.is_staff and not self.is_registrar_user() and not self.is_sponsored_user() and not self.is_organization_user)
-
-    def is_registrar_user(self):
-        """ Is the user a member of a registrar? """
-        return bool(self.registrar_id)
-
-    def is_sponsored_user(self):
-        """ Is the user sponsored by a registrar? """
-        return self.sponsorships.exists()
-
-    def has_registrar_pending(self):
-        """ Has requested creation of registrar """
-        return bool(self.pending_registrar)
-
-    @cached_property
-    def is_organization_user(self):
-        """ Is the user a member of an org? """
-        if self.is_anonymous:
-            return False
-        return self.organizations.exists()
-
-    def is_supported_by_registrar(self):
-        """ Should the user's support requests be forwarded to their registrar?"""
-        if self.is_anonymous:
-            return False
-        return settings.CONTACT_REGISTRARS and \
-               (self.is_organization_user or self.is_sponsored_user())
-
-    ### link permissions ###
-
-    def can_view(self, link):
-        """
-            Not all links are viewable by all users -- some users
-            have privileged access to view private links. For example,
-            a user can view their own private links.
-        """
-        if not link.is_private:
-            return True
-        return self.can_edit(link)
-
-    def can_edit(self, link):
-        """ Link is editable if it is in a folder accessible to this user. """
-        if self.is_anonymous:
-            return False
-        if self.is_staff:
-            return True
-        return Folder.objects.accessible_to(self).filter(links=link).exists()
-
-    def can_delete(self, link):
-        """
-            An archive can be deleted if it is less than 24 hours old-style
-            and it was created by a user or someone in the org.
-        """
-        return not link.user_deleted and not link.is_permanent() and self.can_edit(link)
-
-    def can_toggle_private(self, link):
-        if not self.can_edit(link):
-            return False
-        if link.is_private and not self.is_staff and link.private_reason not in ['user', 'old_policy']:
-            return False
-        return True
-
-    def can_edit_registrar(self, registrar):
-        return self.is_staff or self.registrar == registrar
-
-    def can_edit_organization(self, organization):
-        if self.is_staff:
-            return True
-        elif self.registrar:
-            return self.registrar == organization.registrar
-        else:
-            return self.organizations.filter(pk=organization.pk).exists()
-
-
-    ### subscriptions ###
-
-    def links_remaining_in_period(self, period, limit, unlimited=None):
-        today = timezone.now()
-
-        # default to the value of self.unlimited; allow callers to explicitly override
-        if unlimited is None:
-            unlimited = self.unlimited
-
-        # exclude bonus links, sponsored links and links associated with an org
-        personal_links = Link.objects.filter(organization_id=None, folders__sponsored_by=None).exclude(bonus_link=True)
-
-        if unlimited:
-            # UNLIMITED (paid or sponsored)
-            link_count = float("-inf")
-        elif period == 'once':
-            # TRIAL: all non-org links ever
-            if self.cached_subscription_started:
-                link_count = personal_links.filter(creation_timestamp__range=(self.cached_subscription_started, today), created_by_id=self.id).count()
-            else:
-                link_count = personal_links.filter(created_by_id=self.id, organization_id=None).count()
-        elif period == 'monthly':
-            # MONTHLY RECURRING: links this calendar month (or, for new customers, links this month from the moment you started paying us)
-            if self.cached_subscription_started and \
-               self.cached_subscription_started.year == today.year and \
-               self.cached_subscription_started.month == today.month:
-                link_count = personal_links.filter(creation_timestamp__range=(self.cached_subscription_started, today), created_by_id=self.id, organization_id=None).count()
-            else:
-                link_count = personal_links.filter(creation_timestamp__year=today.year, creation_timestamp__month__gte=today.month, created_by_id=self.id, organization_id=None).count()
-        elif period == 'annually':
-            # ANNUAL RECURRING
-            # if you have a paid subscription, calculate via its expiry date
-            if self.cached_paid_through:
-                link_count = personal_links.filter(creation_timestamp__range=(self.cached_paid_through - relativedelta(years=1), today), created_by_id=self.id, organization_id=None).count()
-            # else, check the last 365 days
-            link_count = personal_links.filter(creation_timestamp__range=(today - relativedelta(years=1), today), created_by_id=self.id, organization_id=None).count()
-        else:
-            raise NotImplementedError("User's link_limit_period not yet handled.")
-        return max(limit - link_count, 0)
-
-    def get_links_remaining(self):
-        """
-            Calculate how many personal links remain.
-            Returns a tuple: (links, applicable period, bonus links)
-        """
-        # Special handling for non-trial users who lack active paid subscriptions:
-        # apply the same rules that are applied to new users
-        if not self.in_trial and not self.nonpaying and self.subscription_status != 'active':
-            return (self.links_remaining_in_period(settings.DEFAULT_CREATE_LIMIT_PERIOD, settings.DEFAULT_CREATE_LIMIT, unlimited=False), settings.DEFAULT_CREATE_LIMIT_PERIOD, self.bonus_links or 0)
-        return (self.links_remaining_in_period(self.link_limit_period, self.link_limit), self.link_limit_period, self.bonus_links or 0)
-
-    def link_creation_allowed(self):
-        links_remaining, _, bonus_links = self.get_links_remaining()
-        return links_remaining > 0 or bonus_links > 0
-
-    def can_view_usage_plan(self):
-        """
-            Should the user be able to see the usage plan page?
-            Special non-paying users should not see the option to make personal purchases.
-            Only authorized users should be able to see a paying registrar's subscription options.
-        """
-        return not self.nonpaying or (self.is_registrar_user() and not self.registrar.nonpaying)
-
-
-    ### merging accounts ###
-
-    def copy_memberships_from_users(self, users):
-        original_orgs = set(self.organizations.all())
-
-        orgs = set()
-        registrars = set()
-        if self.registrar_id:
-            registrars.add(self.registrar_id)
-        else:
-            orgs.update(original_orgs)
-        for user in users:
-            if user.registrar_id:
-                registrars.add(user.registrar_id)
-            else:
-                orgs.update(user.organizations.all())
-
-        if orgs or registrars:
-            assert not (orgs and registrars), f"This set of users includes both org and registrar users: {self.id}, {', '.join([str(user.id) for user in users])}."
-            if registrars:
-                assert len(registrars) == 1, f"This set of users includes registrar users from multiple registrars: {self.id}, {', '.join([str(user.id) for user in users])}."
-                new_registrar_id = registrars.pop()
-                if not self.registrar_id:
-                    self.registrar_id = new_registrar_id
-                    self.prepend_to_notes(f"Added registrar during the merging of accounts: {new_registrar_id}")
-                    self.save(update_fields=['registrar_id', 'notes'])
-            else:
-                if original_orgs != orgs:
-                    self.prepend_to_notes(f"Added organizations during the merging of accounts: {', '.join([str(o.id) for o in orgs - original_orgs])}")
-                    self.save(update_fields=['notes'])
-                    self.organizations.add(*orgs)
-
-    def soft_delete_after_merge_with_user(self, user):
-        original_email = self.email
-
-        self.email = f"merged_users_{self.id}_and_{user.id}@perma.cc"
-        self.is_active = False
-        self.link_count = 0
-
-        self.prepend_to_notes(f"Original email: { original_email }")
-        if self.registrar_id:
-            self.prepend_to_notes(f"Original registrar: { self.registrar_id }")
-            self.registrar_id = None
-        orgs = list(self.organizations.all())
-        if orgs:
-            self.prepend_to_notes(f"Original orgs: {', '.join([str(o.id) for o in orgs])}")
-            self.organizations.remove(*orgs)
-
-        self.save(update_fields=['email', 'is_active', 'link_count', 'notes', 'registrar_id'])
-        return (original_email, self.email)
-
-    def prepend_to_notes(self, message):
-        if self.notes:
-            self.notes = f"{message}\n\n{self.notes}"
-        else:
-            self.notes = message
-
-    def remove_line_from_notes(self, containing):
-        if self.notes:
-            self.notes = re.sub(f"\n*{containing}.*", '', self.notes)
-
-
-class UserOrganizationAffiliation(models.Model):
-    organization = models.ForeignKey(Organization, on_delete=models.CASCADE)
-    user = models.ForeignKey(LinkUser, on_delete=models.CASCADE, db_column='linkuser_id')
-    expires_at = models.DateTimeField(blank=True, null=True)
-
-    class Meta:
-        db_table = 'perma_linkuser_organizations'
-        constraints = [
-            models.UniqueConstraint(fields=['organization', 'user'], name='unique_affiliation'),
-        ]
-        indexes = [
-            models.Index(fields=['expires_at'])
-        ]
-
-
-class ApiKey(models.Model):
-    """
-        Based on tastypie.models: https://github.com/django-tastypie/django-tastypie/blob/master/tastypie/models.py#L35
-    """
-    user = models.OneToOneField(LinkUser, related_name='api_key', on_delete=models.CASCADE)
-    key = models.CharField(max_length=128, blank=True, default='', db_index=True)
-    created = models.DateTimeField(default=timezone.now)
-
-    def __str__(self):
-        return f"{self.key} for {self.user}"
-
-    def save(self, *args, **kwargs):
-        if not self.key:
-            self.key = self.generate_key()
-        return super(ApiKey, self).save(*args, **kwargs)
-
-    def generate_key(self):
-        # Get a random UUID.
-        new_uuid = uuid.uuid4()
-        # Hmac that beast.
-        return hmac.new(new_uuid.bytes, digestmod=hashlib.sha1).hexdigest()
-
-
-# special history tracking for custom user object -- see http://django-simple-history.readthedocs.org/en/latest/reference.html
-simple_history.register(LinkUser)
-
-# This ugly business makes these functions available on logged-out users as well as logged-in,
-# by monkeypatching Django's AnonymousUser object.
-# See https://code.djangoproject.com/ticket/20313
-for func_name in ['can_view', 'can_edit', 'can_delete', 'can_toggle_private', 'is_supported_by_registrar']:
-    setattr(django.contrib.auth.models.AnonymousUser, func_name, getattr(LinkUser, func_name))
-for prop_name in ['is_organization_user']:
-    setattr(django.contrib.auth.models.AnonymousUser, prop_name, getattr(LinkUser, prop_name))
-
-class FolderQuerySet(TreeQuerySet):
-    def user_access_filter(self, user):
-        if user.is_staff:
-            return Q()  # all
-
-        # personal folders
-        folder_list = list(Folder.objects.filter(owned_by=user).values_list('id', flat=True))
-
-        # folders owned by orgs in which the user a member
-        orgs = user.get_orgs()
-        if orgs:
-            folder_list.extend(Folder.objects.filter(organization__in=list(orgs)).values_list('id', flat=True))
-
-        return Q(id__in=folder_list)
-
-    def accessible_to(self, user):
-        return self.filter(self.user_access_filter(user))
-
-
-class Folder(TreeNode):
-    name = models.CharField(max_length=255, null=False, blank=False)
-    creation_timestamp = models.DateTimeField(auto_now_add=True)
-    created_by = models.ForeignKey(settings.AUTH_USER_MODEL, null=True, blank=True, related_name='folders_created', on_delete=models.CASCADE)
-
-    # this may be null if this is the shared folder for a org
-    owned_by = models.ForeignKey(settings.AUTH_USER_MODEL, null=True, blank=True, related_name='folders', on_delete=models.CASCADE)
-
-    # this will be set if this is inside a shared folder
-    organization = models.ForeignKey(Organization, null=True, blank=True, related_name='folders', on_delete=models.CASCADE)
-
-    # true if this is the apex shared folder (not subfolder) for a org, denormalized
-    is_shared_folder = models.BooleanField(default=False)
-
-    # true if this is the apex folder for a user; denormalized
-    is_root_folder = models.BooleanField(default=False)
-
-    # true if this is the apex sponsored folder for a user; denormalized
-    is_sponsored_root_folder = models.BooleanField(default=False)
-    sponsored_by = models.ForeignKey(Registrar, null=True, blank=True, related_name='sponsored_folders', on_delete=models.CASCADE)
-
-    # true if this is a sponsored folder, but the sponsorship is deactivated; denormalized
-    read_only = models.BooleanField(default=False)
-
-    # since a textual representation of the folder's ancestry is included in the folder's API serialization,
-    # keep a cached copy on the model, so we don't have to constantly hit the DB
-    cached_path = models.TextField()
-    cached_has_children = models.BooleanField(default=False)
-
-    # denormalized, a reference to the apex folder in this folder's tree
-    tree_root = models.ForeignKey('self', null=True, on_delete=models.CASCADE)
-
-    objects = FolderQuerySet.as_manager()
-    tracker = FieldTracker()
-
-    def is_leaf_node(self):
-        return not self.cached_has_children
-
-    def get_descendants(self, include_self=False):
-        return self.descendants(
-            include_self=include_self
-        ).tree_filter(
-            tree_root_id=self.tree_root_id
-        )
-
-    def get_ancestors(self, include_self=False):
-        return self.ancestors(include_self=include_self).tree_filter(
-            tree_root_id=self.tree_root_id
-        )
-
-    def delete(self, *args, **kwargs):
-        with transaction.atomic():
-            super().delete(*args, **kwargs)
-            if self.parent_id:
-                Folder.objects.filter(id=self.parent_id).update(
-                    cached_has_children=Exists(
-                        Folder.objects.filter(
-                            parent_id=self.parent_id
-                        )
-                    )
-                )
-
-    def save(self, *args, **kwargs):
-
-        #
-        # Helper methods
-        #
-
-        def get_shared_fields_from_parent(parent):
-            return {
-                "tree_root_id": parent.tree_root_id,
-                "read_only": parent.read_only,
-                "owned_by_id": parent.owned_by_id,
-                "organization_id": parent.organization_id,
-                "sponsored_by_id": parent.sponsored_by_id if not parent.is_sponsored_root_folder else self.sponsored_by_id
-            }
-
-        def set_owner_for_personal_and_sponsored_folders():
-            if self.created_by_id and not self.owned_by_id and not self.organization_id:
-                self.owned_by_id = self.created_by_id
-
-        def get_own_subtree_ids():
-            if self.cached_has_children:
-                return list(
-                    self.get_descendants(
-                        include_self=True
-                    ).tree_filter(
-                        tree_root_id=self.tree_root_id
-                    ).values_list(
-                        'id', flat=True
-                    )
-                )
-            return [self.id]
-
-        def update_tree_root(id, tree_root_id):
-            Folder.objects.filter(id=id).update(
-                tree_root_id=tree_root_id
-            )
-
-        def update_cached_path(ids, tree_root_id):
-            Folder.objects.with_tree_fields().filter(
-                id__in=ids
-            ).update(
-                cached_path=Folder.objects.with_tree_fields().tree_filter(
-                    tree_root_id=tree_root_id
-                ).filter(
-                    id=OuterRef('id')
-                ).extra(
-                    select={"path_string" : "array_to_string((__tree.tree_path), '-')"}
-                ).values_list(
-                    "path_string", flat=True
-                )[:1]
-            )
-
-        def update_parents_cached_has_children(parent_id=None, previous_parent_id=None):
-            if parent_id:
-                Folder.objects.filter(
-                    id=parent_id
-                ).update(
-                    cached_has_children = True
-                )
-            if previous_parent_id:
-                Folder.objects.filter(
-                    id=previous_parent_id
-                ).update(
-                    cached_has_children = Exists(
-                        Folder.objects.exclude(
-                            id=self.id
-                        ).filter(
-                            parent_id=previous_parent_id
-                        )
-                    )
-                )
-
-        #
-        # Save the folder
-        #
-
-        new = not self.pk
-        parent_has_changed = not new and self.tracker.has_changed('parent_id')
-
-        start = time.time()
-        with transaction.atomic():
-
-            if new:
-
-                if self.parent_id:
-
-                    # fetch the parent
-                    parent = Folder.objects.get(id=self.parent_id)
-
-                    # copy shared fields from parent
-                    for field, value in get_shared_fields_from_parent(parent).items():
-                        setattr(self, field, value)
-
-                    # set ownership, if appropriate
-                    set_owner_for_personal_and_sponsored_folders()
-
-                    # simple insert
-                    super().save(*args, **kwargs)
-
-                    # update the cached path, now that the folder has an "id"
-                    update_cached_path([self.id], parent.tree_root_id)
-
-                    # inform the parent it has a new child
-                    update_parents_cached_has_children(parent_id=parent.id)
-
-                else:
-
-                    # set ownership, if appropriate
-                    set_owner_for_personal_and_sponsored_folders()
-
-                    # simple insert
-                    super().save(*args, **kwargs)
-
-                    # update the tree root and cached path, now that the folder has an "id"
-                    update_tree_root(self.id, self.id)
-                    update_cached_path([self.id], self.id)
-
-            elif parent_has_changed:
-
-                # make note of the former parent and the new one
-                parent = Folder.objects.get(id=self.parent_id)
-                previous_parent_id = self.tracker.previous('parent_id')
-
-                # retrieve the ids of this folder and all its descendants, so we can propagate changes.
-                # do it before calling "save", while the database is still in a consistent state
-                # and tree queries work as expected
-                subtree_ids = get_own_subtree_ids()
-
-                # save the change to this folder's parent id
-                super().save(*args, **kwargs)
-
-                # copy shared fields from parent to this folder and all its descendants
-                subtree = Folder.objects.filter(id__in=subtree_ids)
-                subtree.update(**get_shared_fields_from_parent(parent))
-
-                # update the de-normalized reference to owning org on any links in this folder's subtree
-                links = Link.objects.filter(folders__in=subtree_ids)
-                links.update(organization_id=parent.organization_id)
-
-                # if any bonus links got transferred to an org or to a sponsored folder, give users their bonus credit back
-                bonus_links = links.filter(bonus_link=True)
-                if (parent.organization_id or parent.sponsored_by_id) and (any_link := bonus_links.first()):
-                    user = any_link.created_by
-                    count = bonus_links.update(bonus_link=False)
-                    user.bonus_links = F('bonus_links') + count
-                    user.save(update_fields=['bonus_links'])
-
-                # update the cached paths of this folder and all its descendants
-                update_cached_path(subtree_ids, parent.tree_root_id)
-
-                # now that the move is over, inform the parent it has a new child, and inform the previous parent it has one fewer
-                update_parents_cached_has_children(parent_id=parent.id, previous_parent_id=previous_parent_id)
-
-            else:
-                super().save(*args, **kwargs)
-
-        logger.debug(f"Saved {self.id} in {time.time() - start}s")
-
-    class Meta:
-        ordering = ['name', 'id']
-
-    def is_empty(self):
-        return not self.children.exists() and not self.links.exists()
-
-    def __str__(self):
-        return self.name
-
-    @classmethod
-    def format_tree_path(cls, tree_path):
-        return '-'.join([str(i) for i in tree_path])
-
-    def get_path(self):
-        try:
-            return Folder.format_tree_path(self.tree_path)
-        except AttributeError:
-            return Folder.format_tree_path(Folder.objects.with_tree_fields().tree_filter(
-                tree_root_id=self.tree_root_id
-            ).get(id=self.id).tree_path)
-
-    def accessible_to(self, user):
-        # staff can access any folder
-        if user.is_staff:
-            return True
-
-        # private folders (including sponsored folders when viewed by sponsored users)
-        if self.owned_by_id == user.pk:
-            return True
-
-        # sponsored
-        elif self.sponsored_by_id:
-            return self.sponsored_by_id == user.registrar_id
-
-        # shared folders
-        elif self.organization_id:
-            if user.is_registrar_user():
-                # if user is registrar, must be registrar for this org
-                return user.registrar_id == self.organization.registrar_id
-            else:
-                # else, user must belong to this org
-                return user.organizations.filter(pk=self.organization_id).exists()
-
-    @property
-    def sponsorship(self):
-        if self.sponsored_by:
-            return Sponsorship.objects.get(user=self.owned_by, registrar_id=self.sponsored_by)
-
-class LinkQuerySet(QuerySet):
-
-    def user_access_filter(self, user):
-        """
-            User can see/modify a link if they created it or it is in an org folder they belong to.
-            Staff can see/modify all links.
-        """
-        if user.is_staff:
-            return Q()  # all
-
-        # personal links
-        folder_list = list(Folder.objects.filter(owned_by=user).values_list('id', flat=True))
-
-        # links owned by orgs in which the user a member
-        orgs = user.get_orgs()
-        if orgs:
-            folder_list.extend(Folder.objects.filter(organization__in=list(orgs)).values_list('id', flat=True))
-
-        return Q(folders__id__in=folder_list)
-
-    def accessible_to(self, user):
-        return self.filter(self.user_access_filter(user))
-
-    def discoverable(self):
-        return self.filter(Link.DISCOVERABLE_FILTER)
-
-    def successful(self):
-        """ Limit queryset to those where any non-favicon capture succeeded"""
-        return self.filter(
-            captures__in=Capture.objects.filter(Capture.CAN_PLAY_BACK_FILTER)
-        ).distinct()
-
-    def permanent(self):
-        """
-            The required wait period has elapsed, and the user did not delete the Link.
-            It is a permanent part of the collection.
-        """
-        return self.filter(
-            archive_timestamp__lte=timezone.now(),
-            user_deleted=False,
-        )
-
-    def visible_to_memento(self):
-        return self.discoverable().filter(cached_can_play_back=True)
-
-    def visible_to_ia(self):
-        return self.visible_to_memento()
-
-    def ineligible_for_ia(self):
-        return self.exclude(Link.DISCOVERABLE_FILTER, cached_can_play_back=True)
-
-    def ia_upload_pending(self, date_string, limit=100):
-        # Get all Links we think should have been uploaded to IA,
-        # and then filter out the ones that have already been uploaded
-        # to a "daily" item.
-        if date_string > "2022-10-03":
-            # No links created after 2022-10-03 were uploaded to IA as individual Items:
-            # use a simplified query
-            logger.debug("Running simple IA eligibility query.")
-            query = Link.objects.filter(
-                creation_timestamp__date=date_string
-            ).visible_to_ia().filter(
-                internet_archive_files=None
-            )
-        else:
-            logger.debug("Running full IA eligibility query.")
-            query = Link.objects.filter(
-                creation_timestamp__date=date_string
-            ).visible_to_ia().exclude(
-                internet_archive_items__span__isempty=False
-            )
-        if limit:
-            query = query[:limit]
-        return query
-
-
-LinkManager = DeletableManager.from_queryset(LinkQuerySet)
-
-class Link(DeletableModel):
-    """
-    This is the core of the Perma link.
-    """
-    guid = models.CharField(max_length=255, null=False, blank=False, primary_key=True, editable=False)
-    GUID_CHARACTER_SET = "23456789ABCDEFGHJKLMNPQRSTUVWXYZ"
-    replacement_link = models.ForeignKey("Link", blank=True, null=True, help_text="New link to which readers should be forwarded when trying to view this link.", on_delete=models.CASCADE)
-
-    submitted_url = models.URLField(max_length=2100, null=False, blank=False)
-    submitted_url_surt = models.CharField(max_length=2100, null=True, blank=True)
-    creation_timestamp = models.DateTimeField(default=timezone.now, editable=False)
-    submitted_title = models.CharField(max_length=2100, null=False, blank=False)
-    submitted_description = models.CharField(max_length=300, null=True, blank=True)
-    created_by = models.ForeignKey(settings.AUTH_USER_MODEL, null=True, blank=True, related_name='created_links', on_delete=models.CASCADE)
-    organization = models.ForeignKey(Organization, null=True, blank=True, related_name='links', on_delete=models.CASCADE)
-    folders = models.ManyToManyField(Folder, related_name='links', blank=True)
-    notes = models.TextField(blank=True)
-    default_to_screenshot_view = models.BooleanField(default=False, help_text="User defaults to screenshot view.")
-    bonus_link = models.BooleanField(null=True, blank=True)
-
-    captured_by_software = models.CharField(max_length=255, default='perma', db_index=True)
-    captured_by_browser = models.CharField(max_length=255, blank=True, null=True, db_index=True)
-    warc_size = models.IntegerField(blank=True, null=True)
-    wacz_size = models.IntegerField(blank=True, null=True)
-    cached_can_play_back = models.BooleanField(
-        null=True,
-        default=None,
-        db_index=True,
-        help_text="After archive_timestamp, cache whether this link can be played back, for efficiency."
-    )
-
-    is_private = models.BooleanField(default=False)
-    private_reason = models.CharField(max_length=10, blank=True, null=True, choices=(
-        ('domain', 'At the request of the domain owner'),
-        ('meta_perma','Perma-specific robots.txt or meta tag'),
-        ('meta','Generic robots.txt or meta tag'),
-        ('user','At user direction'),
-        ('takedown','At request of content owner'),
-        ('failure','Analysis of meta tags failed'),
-        ('flagged','Contains flagged content')
-    ))
-    is_unlisted = models.BooleanField(default=False)
-
-    archive_timestamp = models.DateTimeField(blank=True, null=True, help_text="Date after which this link is eligible to be copied by the mirror network.")
-    internet_archive_upload_status = models.CharField(max_length=28,
-                                                      default='not_started',
-                                                      choices=(('not_started','not_started'),('completed','completed'),('failed','failed'),('deleted','deleted'), ('deletion_incomplete', 'deletion_incomplete'), ('deletion_required', 'deletion_required'), ('upload_or_reupload_required', 'upload_or_reupload_required')),
-                                                      db_index=True)
-
-    internet_archive_items = models.ManyToManyField(
-        "InternetArchiveItem", through="InternetArchiveFile", related_name="links"
-    )
-
-    objects = LinkManager()
-    tracker = FieldTracker()
-    tags = TaggableManager(through=GenericStringTaggedItem, blank=True)
-
-
-    class Meta:
-        indexes = [
-            models.Index(fields=['user_deleted', 'is_private', 'is_unlisted', 'cached_can_play_back', 'internet_archive_upload_status']),
-            models.Index('user_deleted', TruncDate('creation_timestamp'), 'is_private', 'is_unlisted', 'cached_can_play_back', name="ia_eligible_for_date_idx"),
-            models.Index(fields=['creation_timestamp', 'guid']),
-            models.Index(fields=['-creation_timestamp', 'guid']),
-            models.Index(fields=['submitted_url_surt']),
-            GinIndex(OpClass(Upper('guid'), name='gin_trgm_ops'), name='guid_case_insensitive_idx'),
-        ]
-
-    DISCOVERABLE_FILTER = Q(is_unlisted=False, is_private=False)
-    def is_discoverable(self):
-        return not self.is_private and not self.is_unlisted
-
-    def is_permanent(self):
-        return self.archive_timestamp < timezone.now() and not self.user_deleted
-
-    def has_successful_capture(self):
-        return self.captures.filter(Capture.CAN_PLAY_BACK_FILTER).exists()
-
-    def is_visible_to_memento(self):
-        return self.cached_can_play_back and self.is_discoverable()
-
-    def can_upload_to_internet_archive(self):
-        return self.is_visible_to_memento()
-
-    @cached_property
-    def ia_identifier(self):
-        return settings.INTERNET_ARCHIVE_IDENTIFIER_PREFIX + self.guid
-
-    @classmethod
-    def get_ascii_safe_url(cls, submitted_url):
-        """URL as encoded internally by python requests"""
-        try:
-            # Attempt to quote the URL as well as possible:
-            # - percent encoding
-            # - unicode domains to punycode
-            # - etc.
-            return requests.Request('GET', submitted_url).prepare().url
-        except requests.exceptions.RequestException:
-            # If that fails, just percent encode everything for safety
-            return requests.utils.requote_uri(submitted_url)
-
-    @cached_property
-    def ascii_safe_url(self):
-        """URL as encoded internally by python requests"""
-        return self.get_ascii_safe_url(self.submitted_url)
-
-    @cached_property
-    def url_details(self):
-        return urlparse(self.ascii_safe_url)
-
-    def get_default_title(self):
-        return self.url_details.netloc
-
-    def save(self, *args, **kwargs):
-        # Set a default title if one is missing
-        if not self.submitted_title:
-            self.submitted_title = self.get_default_title()
-
-        initial_folder = kwargs.pop('initial_folder', None)
-
-        if not self.pk:
-            if not self.archive_timestamp:
-                self.archive_timestamp = self.creation_timestamp + settings.ARCHIVE_DELAY
-            if not kwargs.pop("pregenerated_guid", False):
-                # not self.pk => not created yet
-                # only try 100 attempts at finding an unused GUID
-                # (100 attempts should never be necessary, since we'll expand the keyspace long before
-                # there are frequent collisions)
-                r = random.SystemRandom()
-                for i in range(100):
-                    # Generate an 8-character random string like "1A2B3C4D"
-                    guid = ''.join(r.choice(self.GUID_CHARACTER_SET) for _ in range(8))
-
-                    # apply standard formatting (hyphens)
-                    guid = Link.get_canonical_guid(guid)
-
-                    # Avoid GUIDs starting with four letters (in case we need those later)
-                    match = re.search(r'^[A-Z]{4}', guid)
-
-                    if not match and not Link.objects.filter(guid=guid).exists():
-                        break
-                else:
-                    raise Exception("No valid GUID found in 100 attempts.")
-                self.guid = guid
-
-        if not self.submitted_url_surt:
-            self.submitted_url_surt = surt.surt(self.submitted_url)
-
-        if self.is_private and not self.private_reason:
-            self.private_reason = 'user'
-
-        super(Link, self).save(*args, **kwargs)
-
-        if not self.folders.count():
-            if not initial_folder:
-                if self.created_by and self.created_by.root_folder:
-                    initial_folder = self.created_by.root_folder
-            if initial_folder:
-                self.folders.add(initial_folder)
-
-    def __str__(self):
-        return self.guid
-
-    @classmethod
-    def get_canonical_guid(self, guid):
-        """
-        Given a GUID, return the canonical version, with hyphens every 4 chars and all caps.
-        So "a2b3c4d5" becomes "A2B3-C4D5".
-        """
-        # handle legacy 9/10/11-char GUIDs
-        if '-' not in guid and len(guid) >= 9:
-            # handle common typo because legacy URLs start with zero
-            if guid[0] == 'O':
-                guid = guid.replace('O', '0', 1)
-            return guid
-
-        # uppercase and remove non-alphanumerics
-        canonical_guid = re.sub('[^0-9A-Z]+', '', guid.upper())
-
-        # split guid into 4-char chunks, starting from the end
-        guid_parts = [canonical_guid[max(i - 4, 0):i] for i in
-                      range(len(canonical_guid), 0, -4)]
-
-        # stick together parts with '-'
-        return "-".join(reversed(guid_parts))
-
-    def move_to_folder_for_user(self, folder, user):
-        """
-            Move this link to the given folder for the given user.
-        """
-        with transaction.atomic():
-            # Don't let anybody move folders around, until this link is
-            # safely inside its destination folder, lest denormalized
-            # ownership-related fields get out of sync
-            for folder in itertools.chain(self.folders.all(), [folder]):
-                Folder.objects.select_for_update().get(pk=folder.tree_root_id)
-
-            # remove this link from any folders it's in for this user
-            self.folders.remove(*self.folders.accessible_to(user))
-            # add it back to the given folder
-            self.folders.add(folder)
-            if not folder.organization:
-                self.organization = None
-            else:
-                self.organization = folder.organization
-            if self.bonus_link and (folder.organization or folder.sponsored_by):
-                self.bonus_link = False
-                user.bonus_links = F('bonus_links') + 1
-
-            self.save(update_fields=['organization', 'bonus_link'])
-            user.save(update_fields=['bonus_links'])
-
-    def guid_as_path(self):
-        # For a GUID like ABCD-1234, return a path like AB/CD/12.
-        stripped_guid = re.sub('[^0-9A-Za-z]+', '', self.guid)
-        guid_parts = [stripped_guid[i:i + 2] for i in range(0, len(stripped_guid), 2)]
-        return '/'.join(guid_parts[:-1])
-
-    def warc_storage_file(self):
-        return os.path.join(settings.WARC_STORAGE_DIR, self.guid_as_path(), f'{self.guid}.warc.gz')
-
-    def wacz_storage_file(self):
-        return os.path.join(settings.WACZ_STORAGE_DIR, self.guid_as_path(), f'{self.guid}.wacz')
-
-    def warc_to_wacz_conversion_log_file(self):
-        return os.path.join(settings.WACZ_STORAGE_DIR, self.guid_as_path(), f'{self.guid}-conversion-log.json')
-
-    def warc_presigned_url(self):
-        # Specify that warcs should have content-type 'application/gzip' so that archives are fetched correctly by the playback service worker.
-        # (All warcs from before summer 2022 were uploaded with content-type 'application/octet-stream' and content-encoding 'gzip')
-        return storages[settings.WARC_STORAGE].url(self.warc_storage_file(), expire=settings.WARC_PRESIGNED_URL_EXPIRES, parameters={
-                'ResponseContentType': 'application/x-gzip',
-                'ResponseContentEncoding': ''
-        })
-
-    def wacz_presigned_url(self):
-        return storages[settings.WACZ_STORAGE].url(self.wacz_storage_file(), expire=settings.WACZ_PRESIGNED_URL_EXPIRES, parameters={
-            'ResponseContentType': 'application/wacz',
-            'ResponseContentEncoding': ''
-        })
-
-    def warc_presigned_url_relative(self):
-        parsed = urlparse(self.warc_presigned_url())
-        return f'{parsed.path}?{parsed.query}'.lstrip('/')
-
-    def wacz_presigned_url_relative(self):
-        parsed = urlparse(self.wacz_presigned_url())
-        return f'{parsed.path}?{parsed.query}'.lstrip('/')
-
-    def has_wacz_version(self):
-        return bool(self.wacz_size)
-
-    def delete_related_captures(self):
-        Capture.objects.filter(link_id=self.pk).delete()
-
-    def has_capture_job(self):
-        try:
-            self.capture_job
-        except CaptureJob.DoesNotExist:
-            return False
-        return True
-
-    def mark_capturejob_superseded(self):
-        try:
-            job = self.capture_job
-            job.superseded = True
-            job.save()
-        except CaptureJob.DoesNotExist:
-            pass
-
-    @cached_property
-    def is_user_uploaded(self):
-        return self.primary_capture.user_upload
-
-    @cached_property
-    def screenshot_capture(self):
-        return self.captures.filter(role='screenshot').first()
-
-    @cached_property
-    def primary_capture(self):
-        return self.captures.filter(role='primary').first()
-
-    @cached_property
-    def favicon_capture(self):
-        return self.captures.filter(role='favicon').first()
-
-    @cached_property
-    def provenance_summary_capture(self):
-        return self.captures.filter(role='provenance_summary').first()
-
-    @cached_property
-    def pdf_snapshot_capture(self):
-        return self.captures.filter(role='pdf_snapshot').first()
-
-    @cached_property
-    def dom_snapshot_capture(self):
-        return self.captures.filter(role='dom_snapshot').first()
-
-    @cached_property
-    def video_summary_capture(self):
-        return self.captures.filter(role='video_summary').first()
-
-    def get_pages_jsonl(self):
-        if self.can_play_back():
-            jsonl_rows = [
-                {"format": "json-pages-1.0", "id": "pages", "title": "All Pages"}
-            ]
-            ts = str(self.creation_timestamp)
-            if self.provenance_summary_capture:
-                jsonl_rows.append(
-                    {"url": self.provenance_summary_capture.url, "title": "Provenance Summary", "ts": ts}
-                )
-            if self.primary_capture:
-                jsonl_rows.append(
-                    {"url": self.primary_capture.url, "title": f"High-Fidelity Web Capture of {self.ascii_safe_url}", "ts": ts}
-                )
-            if self.screenshot_capture:
-                jsonl_rows.append(
-                    {"url": self.screenshot_capture.url, "title": f"Capture Time Screenshot of {self.ascii_safe_url}", "ts": ts}
-                )
-            if self.dom_snapshot_capture:
-                jsonl_rows.append(
-                    {"url": self.dom_snapshot_capture.url, "title": f"Capture Time DOM snapshot of {self.ascii_safe_url}", "ts": ts}
-                )
-            if self.pdf_snapshot_capture:
-                jsonl_rows.append(
-                    {"url": self.pdf_snapshot_capture.url, "title": f"Capture Time PDF snapshot of {self.ascii_safe_url}", "ts": ts}
-                )
-            if self.video_summary_capture:
-                jsonl_rows.append(
-                    {"url": self.video_summary_capture.url, "title": f"Extracted Video data from: {self.ascii_safe_url}", "ts": ts}
-                )
-            return "\n".join([json.dumps(row) for row in jsonl_rows])
-
-    def write_uploaded_file(self, uploaded_file):
-        """
-            Given a file uploaded by a user, create a Capture record and WACZ.
-        """
-        from api.utils import get_mime_type, mime_type_lookup  # local import to avoid circular import
-
-        # normalize file name to upload.jpg, upload.png, upload.gif, or upload.pdf
-        mime_type = get_mime_type(uploaded_file.name)
-        file_name = f'upload.{mime_type_lookup[mime_type]["new_extension"]}'
-        warc_url = f"file:///{self.guid}/{file_name}"
-
-        upload_capture = Capture(
-            link=self,
-            role='primary',
-            status='success',
-            record_type='resource',
-            user_upload=True,
-            content_type=mime_type,
-            url=warc_url
-        )
-
-        provenance_capture = Capture(
-            link=self,
-            role='provenance_summary',
-            status='success',
-            record_type='resource',
-            user_upload=False,
-            content_type='text/html',
-            url='file:///provenance-summary.html'
-        )
-
-        # make the WACZ
-        self.wacz_size = preserve_perma_wacz(
-            uploaded_file,
-            warc_url,
-            mime_type,
-            self.guid,
-            self.submitted_url,
-            self.submitted_title,
-            self.creation_timestamp,
-            self.wacz_storage_file()
-        )
-        self.warc_size = 0  # necessary?
-
-        self.captured_by_software = 'upload'
-        self.captured_by_browser = None
-        self.save(update_fields=['captured_by_software', 'captured_by_browser', 'warc_size', 'wacz_size'])
-        upload_capture.save()
-        provenance_capture.save()
-
-    def safe_delete_warc(self):
-        old_name = self.warc_storage_file()
-        storage = storages[settings.WARC_STORAGE]
-        if storage.exists(old_name):
-            new_name = old_name.replace('.warc.gz', f'_replaced_{timezone.now().timestamp()}.warc.gz')
-            with storage.open(old_name) as old_file:
-                storage.store_file(old_file, new_name)
-            storage.delete(old_name)
-
-    def safe_delete_wacz(self):
-        old_name = self.wacz_storage_file()
-        storage = storages[settings.WACZ_STORAGE]
-        if storage.exists(old_name):
-            new_name = old_name.replace('.wacz', f'_replaced_{timezone.now().timestamp()}.wacz')
-            with storage.open(old_name) as old_file:
-                storage.store_file(old_file, new_name)
-            storage.delete(old_name)
-        self.wacz_size = 0
-        self.save(update_fields=['wacz_size'])
-
-    @contextmanager
-    def get_warc(self, extract_from_wacz_if_present=True, force_from_wacz=False):
-        if not self.warc_size and not extract_from_wacz_if_present:
-            raise RuntimeError(f'No WARC present for {self.guid}')
-
-        elif self.warc_size and not force_from_wacz:
-            yield storages[settings.WARC_STORAGE].open(self.warc_storage_file(), 'rb')
-
-        elif self.wacz_size:
-            with storages[settings.WACZ_STORAGE].open(self.wacz_storage_file(), 'rb') as wacz_file:
-                yield zipfile.Path(wacz_file, "archive/data.warc.gz").open('rb')
-
-        else:
-            raise RuntimeError(f'No archive present for {self.guid}')
-
-    @contextmanager
-    def get_wacz(self):
-        if not self.wacz_size:
-            raise RuntimeError(f'No WACZ present for {self.guid}')
-        yield storages[settings.WACZ_STORAGE].open(self.wacz_storage_file(), 'rb')
-
-    def accessible_to(self, user):
-        return user.can_edit(self)
-
-    def can_play_back(self):
-        """
-        Reports whether a Perma Link has been successfully captured (or uploaded)
-        and is ready for playback.
-
-        See also /perma/perma_web/static/js/helpers/link.helpers.js
-        """
-        if self.cached_can_play_back is not None:
-            return self.cached_can_play_back
-
-        if self.user_deleted:
-            return False
-
-        successful_metadata = self.has_successful_capture()
-
-        # Early Perma Links and direct uploads do not have CaptureJobs;
-        # if no CaptureJob, judge based on Capture statuses alone;
-        # otherwise, inspect CaptureJob status
-        job = None
-        try:
-            job = self.capture_job
-        except CaptureJob.DoesNotExist:
-            pass
-        if job and not job.superseded and job.status != 'completed':
-            successful_metadata = False
-
-        # Trust our records (the metadata) more than has_warc
-        return successful_metadata
-
-
-class Capture(models.Model):
-    link = models.ForeignKey(Link, null=False, related_name='captures', on_delete=models.CASCADE)
-    role = models.CharField(max_length=18, choices=(
-        ('primary','Primary'),
-        ('screenshot','Screenshot'),
-        ('favicon','Favicon'),
-        ('provenance_summary', 'Provenance Summary'),
-        ('pdf_snapshot', 'PDF Snapshot'),
-        ('dom_snapshot', 'DOM Snapshot'),
-        ('video_summary', 'Video Summary'),
-    ))
-    status = models.CharField(max_length=10, choices=(('pending','pending'),('failed','failed'),('success','success')))
-    url = models.CharField(max_length=2100, blank=True, null=True)
-    record_type = models.CharField(max_length=10, choices=(
-        ('response','WARC Response record -- recorded from web'),
-        ('resource','WARC Resource record -- file without web headers')))
-    content_type = models.CharField(max_length=255, null=False, default='', help_text="HTTP Content-type header.")
-    user_upload = models.BooleanField(default=False, help_text="True if the user uploaded this capture.")
-
-    CAN_PLAY_BACK_FILTER = (Q(role="primary") & Q(status="success")) | (Q(role="screenshot") & Q(status="success"))
-
-    def __str__(self):
-        return f"{self.role} {self.status}"
-
-    def mime_type(self):
-        """
-            Return normalized mime type from content_type.
-            Stuff after semicolon is stripped, type is lowercased, and x- prefix is removed.
-        """
-        return self.content_type.split(";", 1)[0].lower().replace('/x-', '/')
-
-    def use_sandbox(self):
-        """
-            Whether the iframe we use to display this capture should be sandboxed.
-            Answer is yes, unless:
-            a) we're playing back a PDF, which currently can't be sandboxed in Chrome
-            b) the playback will be an on-demand download mediated by an interstitial,
-               because some browsers (for instance, Safari) may block downloads from sandboxed iframes even when `allow-downloads` is present.
-               See https://perma.cc/M36S-ZLVS for `allow-downloads` support on 6/22/22
-        """
-        return not self.mime_type().startswith("application/pdf") and not self.show_interstitial()
-
-    INLINE_TYPES = {
-        'image/jpeg', 
-        'image/gif', 
-        'image/png', 
-        'image/tiff', 
-        'text/html', 
-        'text/plain', 
-        'application/pdf',
-        'application/xhtml', 
-        'application/xhtml+xml', 
-        'video/mp4', 
-        'video/webm', 
-        'video/ogg', 
-        'application/ogg',
-        'audio/ogg',
-        'audio/mpeg',
-        'audio/x-wav',
-        'audio/wav'
-    }
-
-    def show_interstitial(self):
-        """
-            Whether we should show an interstitial view/download button instead of showing the content directly.
-            True unless we recognize the mime type as something that should be shown inline (PDF/HTML/image/audio/video).
-        """
-        return self.mime_type() not in self.INLINE_TYPES
-
-
-def get_default_archive_formats():
-    return ['warc']
-
-class CaptureJob(models.Model):
-    """
-        This class tracks capture jobs for purposes of:
-            (1) sorting the capture queue fairly and
-            (2) reporting status during a capture.
-    """
-    creation_timestamp = models.DateTimeField(auto_now=True, blank=True, null=True, db_index=True)
-    link = models.OneToOneField(Link, related_name='capture_job', null=True, blank=True, on_delete=models.CASCADE)
-    status = models.CharField(max_length=15,
-                              default='invalid',
-                              choices=(('pending','pending'),('in_progress','in_progress'),('completed','completed'),('deleted','deleted'),('failed','failed'),('invalid', 'invalid')),
-                              db_index=True)
-    message = models.TextField(null=True, blank=True) #if we move to postgres, can be a json field
-    human = models.BooleanField(default=False)
-    order = models.FloatField(db_index=True)
-    submitted_url = models.CharField(max_length=2100, blank=True, null=False)
-    created_by = models.ForeignKey(LinkUser, blank=False, null=False, related_name='capture_jobs', on_delete=models.CASCADE)
-    link_batch = models.ForeignKey('LinkBatch', blank=True, null=True, related_name='capture_jobs', on_delete=models.CASCADE)
-    validation_status_code = models.IntegerField(blank=True, null=True)
-    engine = models.CharField(max_length=255,
-                              choices=(('perma', 'perma'), ('scoop-api', 'scoop-api')),
-                              default='scoop-api',
-                              db_index=True)
-
-    archive_formats = ArrayField(
-        models.CharField(
-            max_length=15,
-            choices=(('warc','warc'),('wacz','wacz'))
-        ),
-        default=get_default_archive_formats
-    )
-
-    # reporting
-    attempt = models.SmallIntegerField(default=0)
-    step_count = models.FloatField(default=0)
-    step_description = models.CharField(max_length=255, blank=True, null=True)
-    capture_start_time = models.DateTimeField(blank=True, null=True)
-    capture_end_time = models.DateTimeField(blank=True, null=True)
-    scoop_start_time = models.DateTimeField(blank=True, null=True)
-    scoop_end_time = models.DateTimeField(blank=True, null=True)
-    scoop_logs = JSONField(blank=True, null=True)
-    scoop_job_id = models.CharField(max_length=255, blank=True, null=True, db_index=True)
-    scoop_state = models.CharField(max_length=255, blank=True, null=True, db_index=True)
-
-    superseded = models.BooleanField(default=False, help_text='A user upload has made this CaptureJob irrelevant to the playback of its related Link')
-
-    # settings to allow our tests to draw out race conditions
-    TEST_PAUSE_TIME = 0
-    TEST_ALLOW_RACE = False
-
-    def __str__(self):
-        return f"CaptureJob {self.pk}: {self.link_id}"
-
-    class Meta:
-        indexes = [
-            models.Index(fields=['capture_start_time']),
-        ]
-
-    def save(self, *args, **kwargs):
-
-        # If this job does not have an order yet (just created),
-        # examine all pending jobs to place this one in a fair position in the queue.
-        # "Fair" means round robin: this job will be processed after every other job submitted by this user,
-        # and then after every other user waiting in line has had at least one job done.
-        if not self.order:
-
-            # get all pending jobs, in reverse priority order
-            pending_jobs = CaptureJob.objects.filter(status='pending', human=self.human).order_by('-order').select_related('link')
-            # narrow down to just the jobs that come *after* the most recent job submitted by this user
-            pending_jobs = list(itertools.takewhile(lambda x: x.created_by_id != self.created_by_id, pending_jobs))
-            # flip the list of jobs back around to the order they'll be processed in
-            pending_jobs = list(reversed(pending_jobs))
-
-            # Go through pending jobs until we find two jobs submitted by the same user.
-            # It's not fair for another user to run two jobs after all of ours are done,
-            # so this new job should come right before that user's second job.
-            next_jobs = {}
-            last_job = None
-            for pending_job in pending_jobs:
-                pending_job_created_by_id = pending_job.link.created_by_id
-                if pending_job_created_by_id in next_jobs:
-                    # pending_job is the other user's second job, so this one goes in between that and last_job
-                    self.order = last_job.order + (pending_job.order - last_job.order)/2
-                    break
-                next_jobs[pending_job_created_by_id] = pending_job
-                last_job = pending_job
-
-            # If order isn't set yet, that means we should go last. Find the highest current order and add 1.
-            if not self.order:
-                if pending_jobs:
-                    self.order = pending_jobs[-1].order + 1
-                else:
-                    self.order = (CaptureJob.objects.aggregate(Max('order'))['order__max'] or 0) + 1
-
-        super(CaptureJob, self).save(*args, **kwargs)
-
-    @classmethod
-    def get_next_job(cls, reserve=False):
-        """
-            Return the next job to work on, looking first at the human queue and then at the robot queue.
-
-            If `reserve=True`, mark the returned job with `status=in_progress` and remove from queue so the
-            same job can't be returned twice. Caller must make sure the job is actually processed once returned.
-        """
-
-        # cleanup: mark any captures as deleted where link has been deleted before capture
-        CaptureJob.objects.filter(link__user_deleted=True, status='pending').update(status='deleted')
-
-        while True:
-            next_job = cls.objects.filter(status='pending').order_by('-human', 'order', 'pk').first()
-
-            if reserve and next_job:
-                if cls.TEST_PAUSE_TIME:
-                    time.sleep(cls.TEST_PAUSE_TIME)
-
-                # update the returned job to be in_progress instead of pending, so it won't be returned again
-                # set time using database time, so timeout comparisons will be consistent across worker servers
-                update_count = CaptureJob.objects.filter(
-                    status='pending',
-                    pk=next_job.pk
-                ).update(
-                    status='in_progress',
-                    capture_start_time=Now()
-                )
-
-                # if no rows were updated, another worker claimed this job already -- try again
-                if not update_count and not cls.TEST_ALLOW_RACE:
-                    continue
-
-                # load up-to-date time from database
-                next_job.refresh_from_db()
-
-            return next_job
-
-    def queue_position(self):
-        """
-            Search job_queues to calculate the queue position for this job -- how many pending jobs have to be processed
-            before this one?
-
-            Returns 0 if job is not pending.
-        """
-        if self.status != 'pending':
-            return 0
-
-        queue_position = CaptureJob.objects.filter(status='pending', order__lte=self.order, human=self.human).count()
-        if not self.human:
-            queue_position += CaptureJob.objects.filter(status='pending', human=True).count()
-
-        return queue_position
-
-    def inc_progress(self, inc, description):
-        self.step_count = int(self.step_count) + inc
-        self.step_description = description
-        self.save(update_fields=['step_count', 'step_description'])
-
-    def mark_completed(self, status='completed'):
-        """
-            Record completion time and status for this job.
-        """
-        if status == 'completed' and self.link and self.link.captures.count() == 0:
-            logger.error(f"To investigate: {self.link.guid} has no captures, but was being marked complete")
-            status = 'failed'
-        self.status = status
-        self.capture_end_time = timezone.now()
-        self.save(update_fields=['status', 'capture_end_time', 'message'])
-
-    def mark_failed(self, message):
-        """ Mark job as failed, and record message in format for front-end display. """
-        self.message = json.dumps({api_settings.NON_FIELD_ERRORS_KEY: [message]})
-        self.mark_completed('failed')
-
-    def accessible_to(self, user):
-        return self.link.accessible_to(user)
-
-
-class LinkBatch(models.Model):
-    created_by = models.ForeignKey(LinkUser, blank=False, null=False, related_name='link_batches', on_delete=models.CASCADE)
-    started_on = models.DateTimeField(auto_now=True, blank=False, null=False, db_index=True)
-    target_folder = models.ForeignKey(Folder, blank=False, null=False, on_delete=models.CASCADE)
-    cached_capture_job_count = models.IntegerField(default=0, db_index=True)
-
-    class Meta:
-        verbose_name_plural = "link batches"
-
-    def accessible_to(self, user):
-        return user.is_staff or self.created_by_id == user.pk
-
-    def __str__(self):
-        return f"LinkBatch {self.pk}"
-
-
-#################################
-# Internet Archive related models
-#################################
-
-def get_empty_datetime_range():
-    return DateTimeTZRange(empty=True)
-
-
-class InternetArchiveItem(models.Model):
-    """
-    Archive.org is made up of "items". An item is a logical "thing" that we represent on one web
-    page on archive.org. An item can be considered as a group of files that deserve their own
-    metadata.
-
-    An item is just a directory of files and possibly subdirectories. Every item has at least two
-    files, <identifier>_files.xml and <identifier>_meta.xml. The _meta.xml file is an XML file
-    containing all of the metadata describing the item. The _files.xml file is an XML file
-    containing all of the file-level metadata. There can only be one _meta.xml file and one
-    _files.xml file per item. Alongside these metadata files and the original files uploaded to the
-    item, the item may also contain derivative files automatically generated by archive.org.
-
-    As a rule of thumb, items should:
-    - not be over 100GB
-    - not contain more than 10,000 files.
-
-    (From documentation archived at https://perma.cc/S4WC-64AF)
-
-    Expect one InternetArchiveItem object for every item in the perma_cc Internet Archive collection.
-    """
-
-    # Each item at Internet Archive has an identifier. An identifier is composed
-    # of a unique combination of alphanumeric characters (limited to ASCII), underscores (_),
-    # dashes (-), or periods (.). The first character of an identifier must be alphanumeric
-    # (e.g. it cannot start out with an underscore, dash, or period). The maximum length of
-    # an identifier is 100 characters, but we generally recommend that identifiers be between
-    # 5 and 80 characters in length. Identifiers must be unique across the entirety of Internet
-    # Archive, not simply unique within a single collection.
-    # (From documentation archived at https://perma.cc/AVZ8-FD57)
-    identifier = models.CharField(max_length=100, null=False, blank=False, primary_key=True, editable=False)
-    confirmed_exists = models.BooleanField(
-        default=False,
-        db_index=True,
-        help_text="If false we have attempted to create that item, but have not yet confirmed whether that attempt succeeded; failed attempts are not uncommon."
-    )
-
-    # The best available field reporting when an Item was added to Internet Archive.
-    # (From documentation archived at https://perma.cc/U88M-6R5C)
-    # If the item has been "darked" by IA (e.g. removed after a copyright or content complaint),
-    # this field is no longer returned by the API, so may be unrecoverable.
-    added_date = models.DateTimeField(null=True, blank=True)
-
-    # InternetArchiveItem.objects.filter(span__isempty=True) were created for a single Perma Link
-    # InternetArchiveItem.objects.filter(span__isempty=False) were created for all Perma Links created during this span of time
-    # (https://docs.djangoproject.com/en/4.1/ref/contrib/postgres/fields/#isempty)
-    span = DateTimeRangeField(default=get_empty_datetime_range, help_text="The lower bound is included, and the upper bound excluded. That is, bounds='[)'")
-
-    cached_is_dark = models.BooleanField(default=False, db_index=True)
-    cached_file_count = models.IntegerField(null=True, blank=True, default=None)
-    cached_title = models.TextField(null=True, blank=True, default=None)
-    cached_description = models.TextField(null=True, blank=True, default=None)
-
-    tasks_in_progress = models.IntegerField(default=0, db_index=True, help_text="We have asked Internet Archive to run appx this many tasks for this item and have not yet confirmed that those tasks are complete; derivative tasks not counted.")
-    complete = models.BooleanField(default=False, help_text="Has all the files it ought to have; has no files it ought not have.")
-    last_derived = models.DateTimeField(null=True, blank=True)
-    derive_required = models.BooleanField(default=False)
-
-    class Meta:
-        verbose_name = "Internet Archive Item"
-
-        indexes = [
-            # We would like an index like the below, but expressions aren't supported in this version of Django.
-            # We are adding it via a SQL migration instead. See 0007_auto_20221024_2049.py
-            # models.Index(IsEmpty('span'), 'identifier', name='empty_span_idx'),
-            GistIndex(fields=['span']),
-        ]
-
-    def __str__(self):
-        return f"IA Item {self.identifier}"
-
-    INDIVIDUAL_LINK_IDENTIFIER = '{prefix}{guid}'
-    DAILY_IDENTIFIER = '{prefix}{date_string}'
-
-    @classmethod
-    def datetime(cls, datetime_string):
-        return datetime.strptime(datetime_string, '%Y-%m-%d %H:%M:%S').replace(tzinfo=tz.utc)
-
-    @classmethod
-    def standard_metadata_for_date(cls, date_string):
-        metadata = {
-            "collection": settings.INTERNET_ARCHIVE_COLLECTION,
-            "contributor": "Perma.cc",
-            "mediatype": "web",
-            "date": date_string,
-            "title": f"{date_string} Perma.cc Captures",
-            "description": f"Captures by Perma.cc from {date_string} (one WARC file and XML metadata file per webpage)",
-        }
-        return metadata
-
-    @classmethod
-    def inflight_task_count(cls):
-        return cls.objects.aggregate(Sum('tasks_in_progress'))['tasks_in_progress__sum']
-
-
-class InternetArchiveFile(models.Model):
-    """
-    A file we uploaded to an Internet Archive "Item".
-
-    Items may contain IA-produced, derivative files as well; for instance, cdx files.
-    We do not store objects referencing those derivative files, only original files that we ourselves upload.
-    """
-    link = models.ForeignKey("Link", on_delete=models.DO_NOTHING, related_name='internet_archive_files')
-    item = models.ForeignKey("InternetArchiveItem", on_delete=models.CASCADE, related_name='internet_archive_files')
-
-    status = models.CharField(
-        max_length=19,
-        null=True,
-        blank=True,
-        choices=((s, s) for s in ('upload_attempted', 'upload_submitted', 'confirmed_present', 'deletion_attempted', 'deletion_submitted', 'confirmed_absent')),
-        db_index=True
-    )
-
-    cached_size = models.IntegerField(null=True, blank=True, default=None)
-
-    # Standard metadata fields, defined at https://archive.org/services/docs/api/metadata-schema/index.html#file-metadata-schema
-    cached_title = models.TextField(null=True, blank=True, default=None)
-    cached_comments = models.TextField(null=True, blank=True, default=None)
-    cached_external_identifier = models.TextField(null=True, blank=True, default=None)
-    cached_external_identifier_match_date = models.TextField(null=True, blank=True, default=None)
-    cached_format = models.TextField(null=True, blank=True, default=None)
-
-    # Custom metadata fields
-    cached_submitted_url = models.TextField(null=True, blank=True, default=None)
-    cached_perma_url = models.TextField(null=True, blank=True, default=None)
-
-    tracker = FieldTracker()
-
-    class Meta:
-        verbose_name = "Internet Archive File"
-        unique_together = (("link", "item"),)
-
-    def __str__(self):
-        return f"IA File {self.pk}: {self.item_id} > {self.link_id}"
-
-    WARC_FILENAME = '{guid}.warc.gz'
-
-    @classmethod
-    def guid_from_filename(cls, filename):
-        return filename.split('.')[0]
-
-    @classmethod
-    def standard_metadata_for_link(cls, link):
-        title = f"{link.guid}: {truncatechars(link.submitted_title, 50)}"
-        url = remove_control_characters(link.submitted_url)
-        return {
-            "title": title,
-            "comments": f"Perma.cc archive of {url} created on {link.creation_timestamp}.",
-            "external-identifier": f"urn:X-perma:{link.guid}",
-            "external-identifier-match-date": f"X-perma:{link.creation_timestamp.isoformat()}",
-            "format": "Web ARChive GZ",
-            "submitted_url": url,
-            "perma_url": protocol() + settings.HOST + reverse('single_permalink', args=[link.guid]),
-        }
-
-    def update_from_ia_metadata(self, metadata):
-        """
-        Intentionally does not call save(), so this method can be used with bulk database operations.
-        """
-        self.cached_title = metadata['title']
-        self.cached_comments = metadata['comments']
-        self.cached_external_identifier = metadata['external-identifier']
-        self.cached_external_identifier_match_date = metadata['external-identifier-match-date']
-        self.cached_format = metadata['format']
-        self.cached_submitted_url = metadata['submitted_url']
-        self.cached_perma_url = metadata['perma_url']
-
-    def zero_cached_ia_metadata(self):
-        """
-        Intentionally does not call save(), so this method can be used with bulk database operations.
-        """
-        self.cached_title = None
-        self.cached_comments = None
-        self.cached_external_identifier = None
-        self.cached_external_identifier_match_date = None
-        self.cached_format = None
-        self.cached_submitted_url = None
-        self.cached_perma_url = None
-
