+++ perma_web/perma/models/internet_archive.py
@@ -1,0 +1,199 @@
+from datetime import datetime
+from datetime import timezone as tz
+
+from django.conf import settings
+from django.contrib.postgres.fields import DateTimeRangeField
+from django.contrib.postgres.indexes import GistIndex
+from django.db import models
+from django.db.models import Sum
+from django.template.defaultfilters import truncatechars
+from django.urls import reverse
+from model_utils import FieldTracker
+from psycopg2.extras import DateTimeTZRange
+
+from perma.utils import protocol, remove_control_characters
+
+
+def get_empty_datetime_range():
+    return DateTimeTZRange(empty=True)
+
+
+class InternetArchiveItem(models.Model):
+    """
+    Archive.org is made up of "items". An item is a logical "thing" that we represent on one web
+    page on archive.org. An item can be considered as a group of files that deserve their own
+    metadata.
+
+    An item is just a directory of files and possibly subdirectories. Every item has at least two
+    files, <identifier>_files.xml and <identifier>_meta.xml. The _meta.xml file is an XML file
+    containing all of the metadata describing the item. The _files.xml file is an XML file
+    containing all of the file-level metadata. There can only be one _meta.xml file and one
+    _files.xml file per item. Alongside these metadata files and the original files uploaded to the
+    item, the item may also contain derivative files automatically generated by archive.org.
+
+    As a rule of thumb, items should:
+    - not be over 100GB
+    - not contain more than 10,000 files.
+
+    (From documentation archived at https://perma.cc/S4WC-64AF)
+
+    Expect one InternetArchiveItem object for every item in the perma_cc Internet Archive collection.
+    """
+
+    # Each item at Internet Archive has an identifier. An identifier is composed
+    # of a unique combination of alphanumeric characters (limited to ASCII), underscores (_),
+    # dashes (-), or periods (.). The first character of an identifier must be alphanumeric
+    # (e.g. it cannot start out with an underscore, dash, or period). The maximum length of
+    # an identifier is 100 characters, but we generally recommend that identifiers be between
+    # 5 and 80 characters in length. Identifiers must be unique across the entirety of Internet
+    # Archive, not simply unique within a single collection.
+    # (From documentation archived at https://perma.cc/AVZ8-FD57)
+    identifier = models.CharField(max_length=100, null=False, blank=False, primary_key=True, editable=False)
+    confirmed_exists = models.BooleanField(
+        default=False,
+        db_index=True,
+        help_text="If false we have attempted to create that item, but have not yet confirmed whether that attempt succeeded; failed attempts are not uncommon."
+    )
+
+    # The best available field reporting when an Item was added to Internet Archive.
+    # (From documentation archived at https://perma.cc/U88M-6R5C)
+    # If the item has been "darked" by IA (e.g. removed after a copyright or content complaint),
+    # this field is no longer returned by the API, so may be unrecoverable.
+    added_date = models.DateTimeField(null=True, blank=True)
+
+    # InternetArchiveItem.objects.filter(span__isempty=True) were created for a single Perma Link
+    # InternetArchiveItem.objects.filter(span__isempty=False) were created for all Perma Links created during this span of time
+    # (https://docs.djangoproject.com/en/4.1/ref/contrib/postgres/fields/#isempty)
+    span = DateTimeRangeField(default=get_empty_datetime_range, help_text="The lower bound is included, and the upper bound excluded. That is, bounds='[)'")
+
+    cached_is_dark = models.BooleanField(default=False, db_index=True)
+    cached_file_count = models.IntegerField(null=True, blank=True, default=None)
+    cached_title = models.TextField(null=True, blank=True, default=None)
+    cached_description = models.TextField(null=True, blank=True, default=None)
+
+    tasks_in_progress = models.IntegerField(default=0, db_index=True, help_text="We have asked Internet Archive to run appx this many tasks for this item and have not yet confirmed that those tasks are complete; derivative tasks not counted.")
+    complete = models.BooleanField(default=False, help_text="Has all the files it ought to have; has no files it ought not have.")
+    last_derived = models.DateTimeField(null=True, blank=True)
+    derive_required = models.BooleanField(default=False)
+
+    class Meta:
+        verbose_name = "Internet Archive Item"
+
+        indexes = [
+            # We would like an index like the below, but expressions aren't supported in this version of Django.
+            # We are adding it via a SQL migration instead. See 0007_auto_20221024_2049.py
+            # models.Index(IsEmpty('span'), 'identifier', name='empty_span_idx'),
+            GistIndex(fields=['span']),
+        ]
+
+    def __str__(self):
+        return f"IA Item {self.identifier}"
+
+    INDIVIDUAL_LINK_IDENTIFIER = '{prefix}{guid}'
+    DAILY_IDENTIFIER = '{prefix}{date_string}'
+
+    @classmethod
+    def datetime(cls, datetime_string):
+        return datetime.strptime(datetime_string, '%Y-%m-%d %H:%M:%S').replace(tzinfo=tz.utc)
+
+    @classmethod
+    def standard_metadata_for_date(cls, date_string):
+        metadata = {
+            "collection": settings.INTERNET_ARCHIVE_COLLECTION,
+            "contributor": "Perma.cc",
+            "mediatype": "web",
+            "date": date_string,
+            "title": f"{date_string} Perma.cc Captures",
+            "description": f"Captures by Perma.cc from {date_string} (one WARC file and XML metadata file per webpage)",
+        }
+        return metadata
+
+    @classmethod
+    def inflight_task_count(cls):
+        return cls.objects.aggregate(Sum('tasks_in_progress'))['tasks_in_progress__sum']
+
+
+class InternetArchiveFile(models.Model):
+    """
+    A file we uploaded to an Internet Archive "Item".
+
+    Items may contain IA-produced, derivative files as well; for instance, cdx files.
+    We do not store objects referencing those derivative files, only original files that we ourselves upload.
+    """
+    link = models.ForeignKey("Link", on_delete=models.DO_NOTHING, related_name='internet_archive_files')
+    item = models.ForeignKey("InternetArchiveItem", on_delete=models.CASCADE, related_name='internet_archive_files')
+
+    status = models.CharField(
+        max_length=19,
+        null=True,
+        blank=True,
+        choices=((s, s) for s in ('upload_attempted', 'upload_submitted', 'confirmed_present', 'deletion_attempted', 'deletion_submitted', 'confirmed_absent')),
+        db_index=True
+    )
+
+    cached_size = models.IntegerField(null=True, blank=True, default=None)
+
+    # Standard metadata fields, defined at https://archive.org/services/docs/api/metadata-schema/index.html#file-metadata-schema
+    cached_title = models.TextField(null=True, blank=True, default=None)
+    cached_comments = models.TextField(null=True, blank=True, default=None)
+    cached_external_identifier = models.TextField(null=True, blank=True, default=None)
+    cached_external_identifier_match_date = models.TextField(null=True, blank=True, default=None)
+    cached_format = models.TextField(null=True, blank=True, default=None)
+
+    # Custom metadata fields
+    cached_submitted_url = models.TextField(null=True, blank=True, default=None)
+    cached_perma_url = models.TextField(null=True, blank=True, default=None)
+
+    tracker = FieldTracker()
+
+    class Meta:
+        verbose_name = "Internet Archive File"
+        unique_together = (("link", "item"),)
+
+    def __str__(self):
+        return f"IA File {self.pk}: {self.item_id} > {self.link_id}"
+
+    WARC_FILENAME = '{guid}.warc.gz'
+
+    @classmethod
+    def guid_from_filename(cls, filename):
+        return filename.split('.')[0]
+
+    @classmethod
+    def standard_metadata_for_link(cls, link):
+        title = f"{link.guid}: {truncatechars(link.submitted_title, 50)}"
+        url = remove_control_characters(link.submitted_url)
+        return {
+            "title": title,
+            "comments": f"Perma.cc archive of {url} created on {link.creation_timestamp}.",
+            "external-identifier": f"urn:X-perma:{link.guid}",
+            "external-identifier-match-date": f"X-perma:{link.creation_timestamp.isoformat()}",
+            "format": "Web ARChive GZ",
+            "submitted_url": url,
+            "perma_url": protocol() + settings.HOST + reverse('single_permalink', args=[link.guid]),
+        }
+
+    def update_from_ia_metadata(self, metadata):
+        """
+        Intentionally does not call save(), so this method can be used with bulk database operations.
+        """
+        self.cached_title = metadata['title']
+        self.cached_comments = metadata['comments']
+        self.cached_external_identifier = metadata['external-identifier']
+        self.cached_external_identifier_match_date = metadata['external-identifier-match-date']
+        self.cached_format = metadata['format']
+        self.cached_submitted_url = metadata['submitted_url']
+        self.cached_perma_url = metadata['perma_url']
+
+    def zero_cached_ia_metadata(self):
+        """
+        Intentionally does not call save(), so this method can be used with bulk database operations.
+        """
+        self.cached_title = None
+        self.cached_comments = None
+        self.cached_external_identifier = None
+        self.cached_external_identifier_match_date = None
+        self.cached_format = None
+        self.cached_submitted_url = None
+        self.cached_perma_url = None
+
