+++ perma_web/perma/tests/test_models.py
@@ -27,7 +27,7 @@ import pytest
 # Related to Perma Payments
 #
 
-@patch('perma.models.requests.post', autospec=True)
+@patch('perma.models.base.requests.post', autospec=True)
 def test_get_subscription_none_and_no_network_call_if_nonpaying(post, noncustomers):
     # also verify that their value of in_trial is unchanged by the call
     for noncustomer in noncustomers:
@@ -38,7 +38,7 @@ def test_get_subscription_none_and_no_network_call_if_nonpaying(post, noncustome
         assert post.call_count == 0
 
 
-@patch('perma.models.requests.post', autospec=True)
+@patch('perma.models.base.requests.post', autospec=True)
 def test_get_subscription_raises_on_non_200(post, customers):
     post.return_value.ok = False
     for customer in customers:
@@ -48,8 +48,8 @@ def test_get_subscription_raises_on_non_200(post, customers):
         post.reset_mock()
 
 
-@patch('perma.models.process_perma_payments_transmission', autospec=True)
-@patch('perma.models.requests.post', autospec=True)
+@patch('perma.models.base.process_perma_payments_transmission', autospec=True)
+@patch('perma.models.base.requests.post', autospec=True)
 def test_get_subscription_verifies_transmission_valid(post, process, customers):
     post.return_value.status_code = 200
     post.return_value.json.return_value = sentinel.json
@@ -64,8 +64,8 @@ def test_get_subscription_verifies_transmission_valid(post, process, customers):
         process.reset_mock()
 
 
-@patch('perma.models.process_perma_payments_transmission', autospec=True)
-@patch('perma.models.requests.post', autospec=True)
+@patch('perma.models.base.process_perma_payments_transmission', autospec=True)
+@patch('perma.models.base.requests.post', autospec=True)
 def test_get_subscription_raises_if_unexpected_customer_pk(post, process, customers, spoof_pp_response_wrong_pk):
     post.return_value.status_code = 200
     for customer in customers:
@@ -76,8 +76,8 @@ def test_get_subscription_raises_if_unexpected_customer_pk(post, process, custom
         post.reset_mock()
 
 
-@patch('perma.models.process_perma_payments_transmission', autospec=True)
-@patch('perma.models.requests.post', autospec=True)
+@patch('perma.models.base.process_perma_payments_transmission', autospec=True)
+@patch('perma.models.base.requests.post', autospec=True)
 def test_get_subscription_raises_if_unexpected_registrar_type(post, process, customers, spoof_pp_response_wrong_type):
     post.return_value.status_code = 200
     for customer in customers:
@@ -88,8 +88,8 @@ def test_get_subscription_raises_if_unexpected_registrar_type(post, process, cus
         post.reset_mock()
 
 
-@patch('perma.models.process_perma_payments_transmission', autospec=True)
-@patch('perma.models.requests.post', autospec=True)
+@patch('perma.models.base.process_perma_payments_transmission', autospec=True)
+@patch('perma.models.base.requests.post', autospec=True)
 def test_get_subscription_no_subscription(post, process, customers, spoof_pp_response_no_subscription):
     post.return_value.status_code = 200
     for customer in customers:
@@ -108,8 +108,8 @@ def test_get_subscription_no_subscription(post, process, customers, spoof_pp_res
             post.reset_mock()
 
 
-@patch('perma.models.process_perma_payments_transmission', autospec=True)
-@patch('perma.models.requests.post', autospec=True)
+@patch('perma.models.base.process_perma_payments_transmission', autospec=True)
+@patch('perma.models.base.requests.post', autospec=True)
 def test_get_subscription_no_subscription_purchased_bonus(post, process, customers, spoof_pp_response_no_subscription_two_purchases):
     post.return_value.status_code = 200
     for customer in customers:
@@ -123,8 +123,8 @@ def test_get_subscription_no_subscription_purchased_bonus(post, process, custome
             post.reset_mock()
 
 
-@patch('perma.models.process_perma_payments_transmission', autospec=True)
-@patch('perma.models.requests.post', autospec=True)
+@patch('perma.models.base.process_perma_payments_transmission', autospec=True)
+@patch('perma.models.base.requests.post', autospec=True)
 def test_get_subscription_happy_path_sets_customer_trial_period_to_false(post, process, paying_registrar, paying_user, spoof_pp_response_subscription):
     post.return_value.status_code = 200
     for customer in [paying_registrar, paying_user]:
@@ -145,8 +145,8 @@ def test_get_subscription_happy_path_sets_customer_trial_period_to_false(post, p
         assert customer.cached_subscription_started
 
 
-@patch('perma.models.process_perma_payments_transmission', autospec=True)
-@patch('perma.models.requests.post', autospec=True)
+@patch('perma.models.base.process_perma_payments_transmission', autospec=True)
+@patch('perma.models.base.requests.post', autospec=True)
 def test_get_subscription_happy_path_no_change_pending(post, process, paying_registrar, paying_user, spoof_pp_response_subscription):
     post.return_value.status_code = 200
     for customer in [paying_registrar, paying_user]:
@@ -170,8 +170,8 @@ def test_get_subscription_happy_path_no_change_pending(post, process, paying_reg
             post.reset_mock()
 
 
-@patch('perma.models.process_perma_payments_transmission', autospec=True)
-@patch('perma.models.requests.post', autospec=True)
+@patch('perma.models.base.process_perma_payments_transmission', autospec=True)
+@patch('perma.models.base.requests.post', autospec=True)
 def test_get_subscription_happy_path_with_pending_change(post, process, paying_user, spoof_pp_response_subscription_with_pending_change):
     post.return_value.status_code = 200
     customer = paying_user
@@ -236,7 +236,7 @@ def test_one_time_link_limit_new(user_with_links):
     assert user_with_links.links_remaining_in_period('once', 5) ==  0
 
 
-@patch('perma.models.timezone', autospec=True)
+@patch('perma.models.user.timezone', autospec=True)
 def test_one_time_link_limit_with_midmonth_subscription1(mocked_timezone, user_with_links_this_month_before_the_15th):
     fifteenth_of_month = timezone.now().replace(day=15)
     mocked_timezone.now.return_value = fifteenth_of_month
@@ -252,7 +252,7 @@ def test_one_time_link_limit_with_midmonth_subscription1(mocked_timezone, user_w
     assert mocked_timezone.now.call_count == 2
 
 
-@patch('perma.models.timezone', autospec=True)
+@patch('perma.models.user.timezone', autospec=True)
 def test_one_time_link_limit_with_midmonth_subscription2(mocked_timezone, user_with_links_this_month_before_the_15th):
     fifteenth_of_month = timezone.now().replace(day=15)
     fifth_of_month = timezone.now().replace(day=5)
@@ -275,7 +275,7 @@ def test_monthly_link_limit(user_with_links):
     assert user_with_links.links_remaining_in_period('monthly', 2) == 0
 
 
-@patch('perma.models.timezone', autospec=True)
+@patch('perma.models.user.timezone', autospec=True)
 def test_monthly_link_limit_with_midmonth_subscription1(mocked_timezone, user_with_links_this_month_before_the_15th):
     fifteenth_of_month = timezone.now().replace(day=15)
     mocked_timezone.now.return_value = fifteenth_of_month
@@ -290,7 +290,7 @@ def test_monthly_link_limit_with_midmonth_subscription1(mocked_timezone, user_wi
     assert mocked_timezone.now.call_count == 2
 
 
-@patch('perma.models.timezone', autospec=True)
+@patch('perma.models.user.timezone', autospec=True)
 def test_monthly_link_limit_with_midmonth_subscription2(mocked_timezone, user_with_links_this_month_before_the_15th):
     fifteenth_of_month = timezone.now().replace(day=15)
     fifth_of_month = timezone.now().replace(day=5)
@@ -356,7 +356,7 @@ def test_user_link_creation_denied_if_nonpaying_and_over_limit(get_subscription,
     assert get_links_remaining.call_count == 1
 
 
-@patch('perma.models.subscription_is_active', autospec=True)
+@patch('perma.models.base.subscription_is_active', autospec=True)
 @patch('perma.models.LinkUser.get_subscription', autospec=True)
 def test_user_link_creation_allowed_checks_cached_if_pp_down(get_subscription, is_active, paying_user):
     get_subscription.side_effect = PermaPaymentsCommunicationException
@@ -390,8 +390,8 @@ def test_user_link_creation_denied_if_no_subscription_and_over_limit(get_subscri
 
 
 @patch('perma.models.LinkUser.links_remaining_in_period', autospec=True)
-@patch('perma.models.subscription_is_active', autospec=True)
-@patch('perma.models.subscription_has_problem', autospec=True)
+@patch('perma.models.base.subscription_is_active', autospec=True)
+@patch('perma.models.base.subscription_has_problem', autospec=True)
 @patch('perma.models.LinkUser.get_subscription', autospec=True)
 def test_user_link_creation_disallowed_if_subscription_inactive_and_over_limit(get_subscription, has_problem, is_active, links_remaining_in_period, paying_user):
     get_subscription.return_value = sentinel.subscription
@@ -406,8 +406,8 @@ def test_user_link_creation_disallowed_if_subscription_inactive_and_over_limit(g
 
 
 @patch('perma.models.LinkUser.links_remaining_in_period', autospec=True)
-@patch('perma.models.subscription_is_active', autospec=True)
-@patch('perma.models.subscription_has_problem', autospec=True)
+@patch('perma.models.base.subscription_is_active', autospec=True)
+@patch('perma.models.base.subscription_has_problem', autospec=True)
 @patch('perma.models.LinkUser.get_subscription', autospec=True)
 def test_user_link_creation_allowed_if_subscription_inactive_and_under_limit(get_subscription, has_problem, is_active, links_remaining_in_period, paying_user):
     get_subscription.return_value = sentinel.subscription
@@ -422,7 +422,7 @@ def test_user_link_creation_allowed_if_subscription_inactive_and_under_limit(get
 
 
 @patch('perma.models.LinkUser.links_remaining_in_period', autospec=True)
-@patch('perma.models.subscription_is_active', autospec=True)
+@patch('perma.models.base.subscription_is_active', autospec=True)
 @patch('perma.models.LinkUser.get_subscription', autospec=True)
 def test_user_link_creation_allowed_if_subscription_active_and_under_limit(get_subscription, is_active, links_remaining_in_period, paying_user):
     get_subscription.return_value = sentinel.subscription
@@ -436,7 +436,7 @@ def test_user_link_creation_allowed_if_subscription_active_and_under_limit(get_s
 
 
 @patch('perma.models.LinkUser.links_remaining_in_period', autospec=True)
-@patch('perma.models.subscription_is_active', autospec=True)
+@patch('perma.models.base.subscription_is_active', autospec=True)
 @patch('perma.models.LinkUser.get_subscription', autospec=True)
 def test_user_link_creation_disallowed_if_subscription_active_and_under_limit(get_subscription, is_active, links_remaining_in_period, paying_user):
     get_subscription.return_value = sentinel.subscription
@@ -459,7 +459,7 @@ def test_registrar_link_creation_always_allowed_if_nonpaying(get_subscription, n
     assert get_subscription.call_count == 0
 
 
-@patch('perma.models.subscription_is_active', autospec=True)
+@patch('perma.models.base.subscription_is_active', autospec=True)
 @patch('perma.models.Registrar.get_subscription', autospec=True)
 def test_registrar_link_creation_allowed_checks_cached_if_pp_down(get_subscription, is_active, paying_registrar):
     get_subscription.side_effect = PermaPaymentsCommunicationException
@@ -478,8 +478,8 @@ def test_registrar_link_creation_disallowed_if_no_subscription(get_subscription,
     get_subscription.assert_called_once_with(paying_registrar)
 
 
-@patch('perma.models.subscription_is_active', autospec=True)
-@patch('perma.models.subscription_has_problem', autospec=True)
+@patch('perma.models.base.subscription_is_active', autospec=True)
+@patch('perma.models.base.subscription_has_problem', autospec=True)
 @patch('perma.models.Registrar.get_subscription', autospec=True)
 def test_registrar_link_creation_disallowed_if_subscription_inactive(get_subscription, has_problem, is_active, paying_registrar):
     get_subscription.return_value = sentinel.subscription
@@ -490,7 +490,7 @@ def test_registrar_link_creation_disallowed_if_subscription_inactive(get_subscri
     is_active.assert_called_once_with(sentinel.subscription)
 
 
-@patch('perma.models.subscription_is_active', autospec=True)
+@patch('perma.models.base.subscription_is_active', autospec=True)
 @patch('perma.models.Registrar.get_subscription', autospec=True)
 def test_registrar_link_creation_allowed_if_subscription_active(get_subscription, is_active, paying_registrar):
     get_subscription.return_value = sentinel.subscription
@@ -544,7 +544,7 @@ def test_annotate_tier_monthly_no_subscription_mid_month(customers):
         assert tier['next_payment'] == next_month
 
 
-@patch('perma.models.subscription_is_active', autospec=True)
+@patch('perma.models.base.subscription_is_active', autospec=True)
 def test_annotate_tier_monthly_no_subscription_last_of_month(is_active, customers):
     is_active.return_value = True
 
@@ -603,7 +603,7 @@ def test_annotate_tier_change_disallowed_with_inactive_annual_subscription(custo
             assert tier['type'] == 'unavailable'
 
 
-@patch('perma.models.subscription_is_active', autospec=True)
+@patch('perma.models.base.subscription_is_active', autospec=True)
 def test_annotate_tier_change_disallowed_with_pending_downgrade(is_active, customers):
     is_active.return_value = True
 
@@ -1086,7 +1086,7 @@ def test_subscription_is_active_with_expired_canceled(expired_cancelled_subscrip
 # crediting users for one-time purchases
 #
 
-@patch('perma.models.requests.post', autospec=True)
+@patch('perma.models.base.requests.post', autospec=True)
 def test_credit_for_purchased_links_increments_for_all(post, paying_user, spoof_pp_response_no_subscription_two_purchases):
     post.return_value.ok = True
 
@@ -1097,7 +1097,7 @@ def test_credit_for_purchased_links_increments_for_all(post, paying_user, spoof_
     assert credited == 60
 
 
-@patch('perma.models.requests.post', autospec=True)
+@patch('perma.models.base.requests.post', autospec=True)
 def test_credit_for_purchased_links_reverses_if_acknowledgment_fails(post, paying_user, spoof_pp_response_no_subscription_two_purchases):
     post.return_value.ok = False
 
@@ -1113,10 +1113,10 @@ def test_credit_for_purchased_links_reverses_if_acknowledgment_fails(post, payin
 # Bonus packages info
 #
 
-@patch.object(perma.models, 'datetime')
-@patch('perma.models.prep_for_perma_payments', autospec=True)
+@patch.object(perma.models.base, 'datetime')
+@patch('perma.models.base.prep_for_perma_payments', autospec=True)
 def test_get_bonus_packages(prep, mock_datetime, paying_user):
-    perma.models.datetime.utcnow.return_value = GENESIS
+    perma.models.base.datetime.utcnow.return_value = GENESIS
     prep.return_value.decode.return_value = sentinel.string
 
     packages = paying_user.get_bonus_packages()
